% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_species_control.R
\name{all_species_control}
\alias{all_species_control}
\title{All_species_control}
\usage{
all_species_control(
  data_connection,
  start_year,
  end_year,
  program,
  ocean,
  country_code,
  path_file = NULL
)
}
\arguments{
\item{data_connection}{{\link[base]{list}} expected. Either output of the function {\link[furdeb]{postgresql_dbconnection}}, which must be done before using the all_species_control}

\item{start_year}{{\link[base]{integer}} expected. Starting year for the control.}

\item{end_year}{{\link[base]{integer}} expected. Ending year for the control.}

\item{program}{{\link[base]{character}} expected. Programs to be controlled. Example of the format for a program topiaid: "fr.ird.referential.ps.common.Program#1239832686262#0.31033946454061234"}

\item{ocean}{{\link[base]{character}} expected. Ocean to be controlled. Examples: 'Indian', 'Atlantic'...etc.}

\item{country_code}{{\link[base]{character}} expected. Countries on wich control will be made. Examples: 'FRA', 'MUS'...etc.}

\item{path_file}{{\link[base]{character}} expected. Path to save the final xlsx.}
}
\value{
The function return two xlsx tables.
}
\description{
This function aims to check the coherence of the species caught in a given ocean according to their distribution area. For each inconsistency, it returns a table containing the details of the observation to correct.
}
