% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_activity_sample_inspector.R
\name{check_activity_sample_inspector}
\alias{check_activity_sample_inspector}
\title{Gives inconsistencies between the sample and the activity in terms of presence}
\usage{
check_activity_sample_inspector(dataframe1, dataframe2, output)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_activity_sample_inspector () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_activity_sample_inspector () function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the check_activity_sample_inspector function is to provide a table of data that contains an inconsistency between the sample and the existence of the activity
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  sample_id}}
}
\itemize{
Dataframe 2:
\item{\code{  sample_id}}
\item{\code{  activity_id}}
}
}
\examples{
#Sample 1 is ok,
#Sample 2 is not linked to an activity
dataframe1 <- data.frame(sample_id = c("1", "2"))
dataframe2 <- data.frame(sample_id = c("1"),
                         activity_id = c("1"))
check_activity_sample_inspector(dataframe1, dataframe2, output = "report")
}
