% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_anapo_activity_consistent_inspector.R
\name{check_anapo_activity_consistent_inspector}
\alias{check_anapo_activity_consistent_inspector}
\title{Gives the inconsistencies between the VMS and the presence of activity}
\usage{
check_anapo_activity_consistent_inspector(
  dataframe1,
  dataframe2,
  output,
  vessel_type = c("1", "2", "5", "6", "10")
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_anapo_activity_consistent_inspector function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_anapo_activity_consistent_inspector function.}

\item{output}{{\link[base]{character}} expected.Kind of expected output. You can choose between "message", "report" or "logical".}

\item{vessel_type}{{\link[base]{character}} expected. Default values: c("1", "2", "5", "6", "10"). List of vessel type code to be inspected}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the check_anapo_activity_consistent_inspector function is to provide a table of data that contains an inconsistency between the VMS and the presence of activit
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  vms_id}}
\item{\code{  vms_date}}
\item{\code{  vessel_code}}
\item{\code{  vms_codevessel}}
\item{\code{  vessel_type}}
}
\itemize{
Dataframe 2:
\item{\code{  activity_id}}
\item{\code{  activity_date}}
\item{\code{  vessel_code}}
}
}
\examples{
#VMS 1, 3, 4 and 5 are ok,
#VMS 2 has not linked to an activity
dataframe1 <- data.frame(vms_id = c("1", "2", "3", "4", "5"),
                         vms_date = as.Date(c("2020/01/01", "2020/01/02", "2020/02/02",
                                              "2020/02/02", "2020/02/03")),
                         vessel_code = c("1", "1", "2", "3", "4"),
                         vms_codevessel = c("vessel_1", "vessel_1", "vessel_2", "vessel_2",
                                            "vessel_4"),
                         vessel_type = c("1", "1", "1", "1", "3"))
dataframe2 <- data.frame(activity_id = c("1", "2"),
                         activity_date = as.Date(c("2020/01/01", "2020/02/02")),
                         vessel_code = c("1", "2"))
check_anapo_activity_consistent_inspector(dataframe1, dataframe2, output = "report")
}
