% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_anapo_inspector.R
\name{check_anapo_inspector}
\alias{check_anapo_inspector}
\title{Gives the inconsistencies activity position and VMS position}
\usage{
check_anapo_inspector(
  dataframe1,
  dataframe2,
  dataframe3,
  activity_crs = 4326,
  harbour_crs = 4326,
  vms_crs = 4326,
  output,
  threshold_number_vms = 20,
  threshold_geographical = 10,
  threshold_time = 7200000,
  threshold_score = 0.5,
  buffer_harbour = 11100
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_anapo_inspector () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_anapo_inspector () function.}

\item{dataframe3}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_anapo_inspector () function.}

\item{activity_crs}{{\link[base]{numeric}} expected. Default values: 4326. Coordinate Reference Systems for the position activity}

\item{harbour_crs}{{\link[base]{numeric}} expected. Default values: 4326. Coordinate Reference Systems for the position harbour}

\item{vms_crs}{{\link[base]{numeric}} expected. Default values: 4326. Coordinate Reference Systems for the position VMS}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{threshold_number_vms}{{\link[base]{numeric}} expected. Default values: 20. Minimum number of VMS positions required per day.}

\item{threshold_geographical}{{\link[base]{numeric}} expected. Default values: 10. Maximum valid distance threshold (Nautical miles) between position and nearest VMS point.}

\item{threshold_time}{{\link[base]{numeric}} expected. Default values: 7200000. Maximum valid distance threshold (milliseconds) between position and VMS point.}

\item{threshold_score}{{\link[base]{numeric}} expected. Default values: 0.5. Minimum valid score between position and VMS point.}

\item{buffer_harbour}{{\link[base]{numeric}} expected. Default values: 11100. Buffer to be used for harbour, in meter}
}
\value{
The function returns a {\link[base]{character}} with output is "message", two {\link[base]{data.frame}} with output is "report" (the first without geographical location and the second with geographical location), a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the check_anapo_inspector function is to provide a table of data that contains an inconsistency between activity position and VMS position
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  activity_id}}
\item{\code{  activity_date}}
\item{\code{  activity_time}}
\item{\code{  activity_position}}
\item{\code{  trip_id}}
}
\itemize{
Dataframe 2:
\item{\code{  trip_id}}
\item{\code{  vessel_code}}
\item{\code{  harbour_position_departure}}
\item{\code{  harbour_position_landing}}
}
\itemize{
Dataframe 3:
\item{\code{  vms_date}}
\item{\code{  vms_time}}
\item{\code{  vms_position}}
\item{\code{  vessel_code}}
}
}
\examples{
#Activity 1, 2 and 3 are ok,
#Activity 4 has a number of VMS below the threshold (threshold_number_vms),
#Activity 5 has no position,
#Activity 6 has has a geographical distance above the threshold (threshold_geographical) and
#           a score below the threshold (threshold_score)
dataframe1 <- data.frame(activity_id = c("1", "2", "3", "4", "5", "6"),
                         activity_date = as.Date(c("2020/01/01", "2020/01/12", "2020/01/12",
                                                   "2020/01/13", "2020/01/12", "2020/01/12")),
                         activity_time = c("05:26:01", "10:41:15", "16:41:15", "03:12:34",
                                           "05:56:12", "23:26:47"),
                         activity_position = c("POINT (1 1)", "POINT (0 0)", "POINT (3 0)",
                                               "POINT (4 4)", NA, "POINT (3 0.6)"),
                         trip_id = c("1", "2", "2", "2", "2", "2"))
dataframe2 <- data.frame(trip_id = c("1", "2"),
                         vessel_code = c("1", "1"),
                         harbour_position_departure = c("POINT (1 1.1)", "POINT (3 3)"),
                         harbour_position_landing = c("POINT (3 3)", "POINT (3 3)"))
dataframe3 <- data.frame(vms_date = as.Date(c("2020/01/01", "2020/01/12", "2020/01/12")),
                         vms_time = c("15:26:01", "10:55:15", "22:32:17"),
                         vms_position = c("POINT (4 4)", "POINT (0 0.1)", "POINT (3 0.3)"),
                         vessel_code = c("1", "1", "1"))
check_anapo_inspector(dataframe1,dataframe2, dataframe3, output = "report",threshold_number_vms = 1)
}
