% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fishing_time_inspector.R
\name{check_fishing_time_inspector}
\alias{check_fishing_time_inspector}
\title{Gives the inconsistencies between the sum of the fishing times indicated for the route and the one indicated for the trip}
\usage{
check_fishing_time_inspector(dataframe1, dataframe2, output)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_fishing_time_inspector () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_fishing_time_inspector () function.}

\item{output}{{\link[base]{character}} expected.Kind of expected output. You can choose between "message", "report" or "logical".}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the check_fishing_time_inspector function is to provide a table of data that contains an inconsistency between the sum of the fishing times indicated for the route and the one indicated for the trip
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  trip_id}}
\item{\code{  trip_fishingtime}}
}
\itemize{
Dataframe 2:
\item{\code{  route_id}}
\item{\code{  trip_id}}
\item{\code{  route_fishingtime}}
}
}
\examples{
#Trip 1 and 5 are ok,
#Trip 2 has different fishing times,
#Trip 3 has no trip_fishingtime,
#Trip 4 has no route_fishingtime
dataframe1 <- data.frame(trip_id = c("1", "2", "3", "4", "5"),
                         trip_fishingtime = c(10, 15, NA, 8, 0))
dataframe2 <- data.frame(route_id = c("1", "2", "3", "4", "5", '6'),
                         trip_id = c("1", "1", "2", "2", "3", "5"),
                         route_fishingtime = c(4, 6, 10, 6, 14, 0))
check_fishing_time_inspector(dataframe1, dataframe2, output = "report")
}
