% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_harbour_inspector.R
\name{check_harbour_inspector}
\alias{check_harbour_inspector}
\title{Gives the inconsistencies between the harbour of landing of the previous trip and the harbour of departure of the current trip}
\usage{
check_harbour_inspector(dataframe1, output)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_harbour_inspector () function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the check_harbour_inspector function is to provide a table of data that contains an inconsistency with the harbour of landing of the previous trip and the harbour of departure of the current trip
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  trip_id}}
\item{\code{  harbour_id_departure}}
\item{\code{  harbour_label_departure}}
\item{\code{  trip_previous_id}}
\item{\code{  harbour_id_landing_trip_previous}}
\item{\code{  harbour_label_landing_trip_previous}}
}
}
\examples{
#Trip 1 and 2 are ok,
#Trip 3 has different harbour,
#Trip 4 has no harbour departure,
#Trip 5 has no harbour landing for trip previous
dataframe1 <- data.frame(trip_id = c("1", "2", "3", "4", "5"),
                         harbour_id_departure = c("1", "2", "3", NA, "1"),
                         harbour_label_departure = c("Harbour_1", "Harbour_2", "Harbour_3", NA,
                                                     "Harbour_1"),
                         trip_previous_id = c(NA, "1", "2", NA, "6"),
                         harbour_id_landing_trip_previous = c(NA, "2", "2", NA, NA),
                         harbour_label_landing_trip_previous = c(NA, "Harbour_2", "Harbour_2", NA,
                                                                 NA))
check_harbour_inspector(dataframe1, output = "report")
}
