% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_landing_total_weight_inspector.R
\name{check_landing_total_weight_inspector}
\alias{check_landing_total_weight_inspector}
\title{Gives the inconsistencies between the total weight landed during the trip for the canneries and the sum of the weights of each landing for the canneries linked to the trip}
\usage{
check_landing_total_weight_inspector(
  dataframe1,
  dataframe2,
  output,
  epsilon = 0.01
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_landing_total_weight_inspector () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_landing_total_weight_inspector () function.}

\item{output}{{\link[base]{character}} expected.Kind of expected output. You can choose between "message", "report" or "logical".}

\item{epsilon}{{\link[base]{numeric}} expected, default : 0.01. Gives the threshold at which the difference is considered too large.}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the check_landing_total_weight_inspector function is to provide a table of data that contains an inconsistency between the sum of the weights of each landing for the canneries and the one indicated for the trip
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  trip_id}}
\item{\code{  trip_landingtotalweight}}
}
\itemize{
Dataframe 2:
\item{\code{  landing_id}}
\item{\code{  landing_weight}}
\item{\code{  trip_id}}
}
}
\examples{
#Trip 1 and 4 are ok,
#Trip 2 has different landing weight,
#Trip 3 has no trip_landingtotalweight,
#Trip 5 has no landing_weight
dataframe1 <- data.frame(trip_id = c("1", "2", "3", "4", "5"),
                         trip_landingtotalweight = c(10, 15, NA, 0, 4))
dataframe2 <- data.frame(landing_id = c("1", "2", "3", "4", "5", "6"),
                         landing_weight = c(4, 6, 10, 6, 2, NA),
                         trip_id = c("1", "1", "2", "2", "3", "4"))
check_landing_total_weight_inspector(dataframe1, dataframe2, output = "report")
}
