% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_little_big_inspector.R
\name{check_little_big_inspector}
\alias{check_little_big_inspector}
\title{Gives the inconsistencies between the percentage of little and big fish sampled}
\usage{
check_little_big_inspector(
  dataframe1,
  dataframe2,
  dataframe3,
  output,
  species_big = c("YFT", "YFT", "BET", "BET", "ALB", "ALB"),
  size_measure_type_big = c("PD1", "FL", "PD1", "FL", "PD1", "FL"),
  threshold_size_big = c(24, 80, 25, 77, 23.5, 78),
  size_measure_type = c("FL", "PD1"),
  threshold = 0.9
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_little_big_inspector() function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_little_big_inspector() function.}

\item{dataframe3}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_little_big_inspector() function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{species_big}{{\link[base]{character}} expected. Default values:  c("YFT", "YFT", "BET", "BET", "ALB", "ALB"). Vector of the species. First criterion for identifying big fish (other values are small fish)}

\item{size_measure_type_big}{{\link[base]{character}} expected. Default values: c("PD1", "FL", "PD1", "FL", "PD1", "FL"). Vector of the size measure type. Second criterion for identifying big fish (other values are small fish)}

\item{threshold_size_big}{{\link[base]{numeric}} expected. Default values: c(24, 80, 25, 77, 23.5, 78). Vector for defining the lower or equal for threshold size measurement. Third criterion for identifying big fish (other values are small fish)}

\item{size_measure_type}{{\link[base]{character}} expected. Default values: c("FL", "PD1"). Vector with the preferred type of size measurement for small fish and then for big fish}

\item{threshold}{{\link[base]{numeric}} expected. Default values: 0.9. Threshold for percentage of small or big fish}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the check_little_big_inspector function is to provide a table of data that contains an inconsistency between the percentage of little and big fish sampled
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  sample_id}}
\item{\code{  sample_smallsweight}}
\item{\code{  sample_bigsweight}}
\item{\code{  sample_totalweight}}
}
\itemize{
Dataframe 2:
\item{\code{  samplespecies_id}}
\item{\code{  species_fao_code}}
\item{\code{  sizemeasuretype_code}}
\item{\code{  sample_id}}
}
\itemize{
Dataframe 3:
\item{\code{  samplespeciesmeasure_id}}
\item{\code{  samplespeciesmeasure_sizeclass}}
\item{\code{  samplespeciesmeasure_count}}
\item{\code{  samplespecies_id}}
}
}
\examples{
#Sample 1, 2, 4, 5, 7, 8 and 9 are ok,
#Sample 3 has a percentage of big below the threshold,
#Sample 6 has a percentage of little below the threshold,
#Sample 10 has a percentage of little below the threshold,
#Sample 11 has a percentage of little below the threshold,
#Sample 12 has a percentage of big below the threshold
dataframe1 <- data.frame(sample_id = c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11",
                                       "12"),
                         sample_smallsweight = c(10, 20, 1, 30, 3, 7, 4, 12, 0, 0, 3, 0),
                         sample_bigsweight = c(NA, 2, 9, 3, 5, 4, 13, 5, 0, NA, 0, 4),
                         sample_totalweight = c(NA, NA, NA, 33, NA, NA, 7, 2, 0, 5, 0, 0))
dataframe2 <- data.frame(samplespecies_id = c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10",
                                              "11", "12"),
                         species_fao_code = c("YFT", "BET", "BET", "YFT", "ALB", "YFT", "YFT",
                                              "BET", "YFT", "ALB", "YFT", "YFT"),
                         sizemeasuretype_code = c("PD1", "PD1", "FL","PD1", "FL", "FL", "PD1",
                                                  "FL", "PD1", "PD1", "PD1", "PD1"),
                         sample_id = c("1", "2", "2", "3", "4", "5", "5", "6", "6", "8", "9", "10"))
dataframe3 <- data.frame(samplespeciesmeasure_id = c("1", "2", "3", "4", "5", "6", "7", "8", "9",
                                                     "10", "11", "12"),
                         samplespeciesmeasure_sizeclass = c(20, 26, 70, 20, 10, 45, 36, 30, 32, 24,
                                                            13, 13),
                         samplespeciesmeasure_count = c(5, 1, 9, 8, 10, 25, 2, 16, 4, 3, 6, 6),
                         samplespecies_id = c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10",
                                              "11", "12"))
check_little_big_inspector(dataframe1, dataframe2, dataframe3, output = "report")
}
