% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_raising_factor_inspector.R
\name{check_raising_factor_inspector}
\alias{check_raising_factor_inspector}
\title{Gives the inconsistencies between RF1 and threshold values}
\usage{
check_raising_factor_inspector(
  dataframe1,
  dataframe2,
  dataframe3,
  dataframe4,
  output,
  country_species = list(`1` = c("TUN", "ALB", "YFT", "BET", "SKJ"), `4` = c("LOT",
    "TUN", "ALB", "YFT", "BET", "SKJ", "LTA", "FRI", "BLF", "RAV*", "KAW", "FRZ", "BLT")),
  speciesfate = "6",
  vesselactivity = c("23", "25", "27", "29"),
  threshold = c(0.9, 1.1)
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_raising_factor_inspector () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_raising_factor_inspector () function.}

\item{dataframe3}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_raising_factor_inspector () function.}

\item{dataframe4}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_raising_factor_inspector () function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{country_species}{{\link[base]{character}} expected. Default values: list("1" = c("TUN", "ALB", "YFT", "BET", "SKJ"), "4" = c("LOT", "TUN", "ALB", "YFT", "BET", "SKJ", "LTA", "FRI", "BLF", "RAV*", "KAW", "FRZ", "BLT")). list of the inventory of species (FAO code) used to calculate catch weight in RF1 by country (country code).}

\item{speciesfate}{{\link[base]{character}} expected. Default values: "6". Vector of inventory of fate used to calculate catch weight in RF1.}

\item{vesselactivity}{{\link[base]{character}} expected. Default values: c("23", "25", "27", "29"). Vector of inventory of vessel activity NOT used to calculate catch weight in RF1.}

\item{threshold}{{\link[base]{numeric}} expected. Default values: 0.9 and 1.1. Vector containing the lower and upper acceptable threshold for RF1.}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the check_raising_factor_inspector function is to provide a table of data that contains an inconsistency with the RF1 and the valid threshold (Default : 0.9 ; 1.1)
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  trip_id}}
Dataframe 2:
\item{\code{  catch_id}}
\item{\code{  catch_weight}}
\item{\code{  speciesfate_code}}
\item{\code{  species_fao_code}}
\item{\code{  vesselactivity_code}}
\item{\code{  trip_id}}
Dataframe 3:
\item{\code{  landing_id}}
\item{\code{  landing_weight}}
\item{\code{  species_fao_code}}
\item{\code{  trip_id}}
Dataframe 4:
\item{\code{  trip_id}}
\item{\code{  trip_end_full_trip_id}}
\item{\code{  vessel_id}}
\item{\code{  country_fleetcountry}}
}
}
