% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sample_without_species_inspector.R
\name{check_sample_without_species_inspector}
\alias{check_sample_without_species_inspector}
\title{Gives inconsistencies between the sample and the species sampled in terms of presence}
\usage{
check_sample_without_species_inspector(dataframe1, dataframe2, output)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_sample_without_species_inspector() function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_sample_without_species_inspector() function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the check_sample_without_species_inspector function is to provide a table of data that contains an inconsistency between the sample and the presence of species
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  sample_id}}
}
\itemize{
Dataframe 2:
\item{\code{  samplespecies_id}}
\item{\code{  sample_id}}
}
}
\examples{
#Sample 1 is ok,
#Sample 2 has no sample species
dataframe1 <- data.frame(sample_id = c("1", "2"))
dataframe2 <- data.frame(samplespecies_id = c("1", "2"),
                         sample_id = c("1", "1"))
check_sample_without_species_inspector(dataframe1, dataframe2, output = "report")
}
