% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_species_catch_ocean.R
\name{check_species_catch_ocean}
\alias{check_species_catch_ocean}
\title{Gives the inconsistencies between the species fished and the ocean reported for the trip}
\usage{
check_species_catch_ocean(data_connection, type_select, select, output)
}
\arguments{
\item{data_connection}{{\link[base]{list}} expected. Either output of the function {\link[furdeb]{postgresql_dbconnection}}, which must be done before using the check_species_catch_ocean function or the two data sets used (the first one is the species-ocean associations and the second one is the species caught with the associated ocean)}

\item{type_select}{{\link[base]{character}} expected. Choose the study unit, you can choose between: "trip" or "year".}

\item{select}{{\link[base]{character}} expected with type_select is "trip". Then give unique identifier for the trip. {\link[base]{numeric}} expected with type_select is "year".Then give year for the trip.}

\item{output}{{\link[base]{character}} expected.Kind of expected output. You can choose between "message", "report" or "logical".}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the check_species_catch_ocean function is to provide a table of data that contains an inconsistency between the species caught during the trip and the ocean reported for the trip with the ocean for which the species are associated
}
