% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_specimen_identifier_typing.R
\name{check_specimen_identifier_typing}
\alias{check_specimen_identifier_typing}
\alias{check_specimen_typing_identifier}
\title{Fish identifier checking in data specimen}
\usage{
check_specimen_typing_identifier(data)
}
\arguments{
\item{data}{Object of type \code{\link[base]{data.frame}} expected.}
}
\description{
Fish identifier checking in data specimen
}
