% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_specimen_lwr.R
\name{check_specimen_lwr}
\alias{check_specimen_lwr}
\title{Fork length checking in data specimen}
\usage{
check_specimen_lwr(data, threshold = 5)
}
\arguments{
\item{data}{Object of type \code{\link[base]{data.frame}} expected.}

\item{threshold}{Object of type \code{\link[base]{numeric}} expected.}
}
\description{
Fork length checking in data specimen
}
