% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_weight_sample_inspector.R
\name{check_weight_sample_inspector}
\alias{check_weight_sample_inspector}
\title{Gives the inconsistencies between the sample weight (m10 and p10) and the global sample weight}
\usage{
check_weight_sample_inspector(dataframe1, output)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_weight_sample_inspector () function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the check_weight_sample_inspector  function is to provide a table of data that contains an inconsistency between the sample weight (m10 and p10) and the global weight
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  sample_id}}
\item{\code{  sample_smallsweight}}
\item{\code{  sample_bigsweight}}
\item{\code{  sample_totalweight}}
}
}
\examples{
#Sample 1, and 2 are ok,
#Sample 3 has total weight and smalls weight,
#Sample 4 has not total weight nor smalls weight nor bigs weight
dataframe1 <- data.frame(sample_id = c("1", "2", "3", "4"),
                         sample_smallsweight = c(10, NA, 12, NA),
                         sample_bigsweight = c(50, NA, NA, 0),
                         sample_totalweight = c(NA, 9, 5, 0))
check_weight_sample_inspector(dataframe1, output = "report")
}
