% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation_references.R
\name{creation_references}
\alias{creation_references}
\title{Creation references}
\usage{
creation_references(path_references, output_path)
}
\arguments{
\item{path_references}{Unique object of type \code{\link[base]{character}} expected. A préciser...}

\item{output_path}{Chemin d'enregistrement des tables de references}
}
\description{
Creation references
}
