% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fate_by_species_group_control.R
\name{fate_by_species_group_control}
\alias{fate_by_species_group_control}
\title{Fate by species control}
\usage{
fate_by_species_group_control(
  data_connection,
  start_year,
  end_year,
  program,
  ocean,
  country_code,
  path_file = NULL,
  export_table
)
}
\arguments{
\item{data_connection}{{\link[base]{list}} expected. Either output of the function {\link[furdeb]{postgresql_dbconnection}}, which must be done before using the fate_control.}

\item{start_year}{{\link[base]{integer}} expected. Starting year for the control.}

\item{end_year}{{\link[base]{integer}} expected. Ending year for the control.}

\item{program}{{\link[base]{character}} expected. Programs to be controlled.}

\item{ocean}{{\link[base]{character}} expected. Ocean to be controlled.}

\item{country_code}{{\link[base]{character}} expected. Countries for which control will be made.}

\item{path_file}{{\link[base]{character}} expected. Path to save the final xlsx with all the informations for the correction.}

\item{export_table}{{\link[base]{logical}} expected. Save or not the xlsx tables.}
}
\value{
The function returns two xlsx tables, one for the catch and one for the samples.
}
\description{
Identifies in the observer data all the species with an inconsistent fate according to their species group.
}
\note{
Version 1.0
}
\author{
Chloé Tellier, Philippe S. Sabarros
}
