% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_catch_fate_unknown_control.R
\name{function_catch_fate_unknown_control}
\alias{function_catch_fate_unknown_control}
\title{Catch fate unknown control}
\usage{
function_catch_fate_unknown_control(catch)
}
\arguments{
\item{catch}{{\link[base]{data.frame}} expected. All catches during the time range selected.}
}
\value{
The function returns one {\link[base]{data.frame}}.
}
\description{
Identifies in the observer data catches where fate code 9 (other) or 11 (discarded unknown) were used.
}
\details{
The input dataframes must contain all these columns for the function to work :
\itemize{
Dataframe catch:
\item{\code{  catch_id}}
\item{\code{  speciesfate_code}}
}
}
\note{
Version 1.0
}
\author{
Chloé Tellier, Esther Mollier, Philippe S. Sabarros
}
\examples{
#Catch 1 is ok: the catch is discarded alive,
#Catch 2 is ok: the catch is retained on board,
#Catch 3 is not ok: the catch has an unknown fate,
#Catch 3 is not ok: the catch is discarded but with an unknown status.
catch <- data.frame(catch_id = c("1", "2", "3", "4"),
                    speciesfate_code = c("4", "15", "9", "11"))
function_catch_fate_unknown_control(catch)
}
