% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_reason_for_discard_unknown_control.R
\name{function_reason_for_discard_unknown_control}
\alias{function_reason_for_discard_unknown_control}
\title{Reason for discard unknown control}
\usage{
function_reason_for_discard_unknown_control(catch)
}
\arguments{
\item{catch}{{\link[base]{data.frame}} expected. All catches during the time range selected.}
}
\value{
The function returns one {\link[base]{data.frame}}.
}
\description{
Identifies in the observer data cases where the reason for discard is 99 or NA.
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe catch:
\item{\code{  catch_id}}
\item{\code{  speciesfate_code}}
\item{\code{  reasonfordiscard_code}}
}
}
\note{
Version 1.0
}
\author{
Chloé Tellier, Philippe S. Sabarros
}
\examples{
#Catch 1 is ok: the catch is not discarded,
#Catch 2 is ok: the catch is discarded and has a reason for discard,
#Catch 3 is not ok: the catch is discarded and has a reason for discard = NA,
#Catch 3 is not ok: the catch is discarded and does not have a reason for discard.
catch <- data.frame(catch_id = c("1", "2", "3", "4"),
                    speciesfate_code = c("6", "4", "5", "14"),
                    reasonfordiscard_code = c(NA, "1", NA, "99"))
function_reason_for_discard_unknown_control(catch)
}
