% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbook_buoy_code_control.R
\name{logbook_buoy_code_control}
\alias{logbook_buoy_code_control}
\title{Identification of an inconsistent buoy code and buoy type naming convention}
\usage{
logbook_buoy_code_control(dataframe1, output)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_buoy_code_control.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical".
}
\description{
The purpose of the logbook_buoy_code_control function is to provide a table of data that contains a incoherent buoy code and buoy type naming convention
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  transmittingbuoy_id}}
\item{\code{  transmittingbuoy_code}}
\item{\code{  transmittingbuoytype_regex}}
}
}
\examples{
#Buoy 1 is ok,
#Buoy 2 has start by 0
#Buoy 3 has a 7 digit identifier.
dataframe1 <- data.frame(transmittingbuoy_id = c("1", "2", "3", "4"),
                         transmittingbuoy_code = c("111111", "011111", "1234567", NA),
                         transmittingbuoytype_regex = c("[0-9]{6}", "[0-9]{6}", "[0-9]{6}",
                                                        "[0-9]{6}"))
logbook_buoy_code_control(dataframe1, output = "report")
}
