% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbook_buoy_owner_control.R
\name{logbook_buoy_owner_control}
\alias{logbook_buoy_owner_control}
\title{Identification of an inconsistent buoy whose owner is missing}
\usage{
logbook_buoy_owner_control(dataframe1, output)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_buoy_owner_control.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical".
}
\description{
The purpose of the logbook_buoy_owner_control function is to provide a table of data that contains a incoherent buoy whose owner is missing
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  transmittingbuoy_id}}
\item{\code{  transmittingbuoyownership_id}}
}
}
\examples{
#Buoy 1 is ok,
#Buoy 2 has owner missing
dataframe1 <- data.frame(transmittingbuoy_id = c("1", "2"),
                         transmittingbuoyownership_id = c("1", NA))
logbook_buoy_owner_control(dataframe1, output = "report")
}
