% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbook_buoy_time_control.R
\name{logbook_buoy_time_control}
\alias{logbook_buoy_time_control}
\title{Gives the inconsistencies between the time elapsed for an operation on the same buoy}
\usage{
logbook_buoy_time_control(dataframe1, dataframe2, output, threshold_time = 60)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_buoy_time_control () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_buoy_time_control () function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{threshold_time}{{\link[base]{numeric}} expected. Default values: 60 Maximum valid time threshold (minutes) between operation on the same buoy.}
}
\value{
The function returns a {\link[base]{character}} with output is "message", two {\link[base]{data.frame}} with output is "report" (the first without geographical location and the second with geographical location), a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the logbook_buoy_time_control function is to provide a table of data that contains an inconsistency between the time elapsed for an operation on a floating object with the same buoy
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  transmittingbuoy_id}}
\item{\code{  transmittingbuoy_code}}
\item{\code{  transmittingbuoytype_id}}
\item{\code{  activity_id}}
}
\itemize{
Dataframe 2:
\item{\code{  activity_id}}
\item{\code{  activity_date}}
\item{\code{  activity_time}}
\item{\code{  trip_id}}
}
}
\examples{
#Buoy 1, 2, 3, 4 and 5 are ok,
#Buoy 6 is followed by an operation on the same object within a time interval shorter than
# the threshold
#Buoy 7 is preceded by an operation on the same object within a time interval shorter than
# the threshold
#Buoy 8 is linked to an activity that involves two operations on the same object
#Buoy 9 is linked to an activity that involves two operations on the same object
dataframe1 <- data.frame(transmittingbuoy_id = c("1", "2", "3", "4", "5", "6", "7", "8", "9"),
                         transmittingbuoy_code = c("1", "1", "1", "1", "2", "3", "3", "4", "4"),
                         transmittingbuoytype_id = c("1", "2", "2", "2", "2", "1", "1", "2", "2"),
                         activity_id = c("1", "2", "3", "4", "5", "6", "7", "8", "8"))
dataframe2 <- data.frame(activity_id = c("1", "2", "3", "4", "5", "6", "7", "8"),
                         activity_date = as.Date(c("2020/01/01", "2020/01/01", "2020/01/01",
                                                   "2020/01/01", "2020/01/01", "2020/01/02",
                                                   "2020/01/02", "2020/01/03")),
                         activity_time = c("15:26:01", "15:36:01", "17:49:00", "18:30:00",
                                           "18:31:00", "09:26:01", "09:42:01", "21:35:01"),
                         trip_id = c("1", "1", "1", "2", "2", "2", "2", "2"))
logbook_buoy_time_control(dataframe1, dataframe2, output = "report")
}
