% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/logbook_category_species_forbidden_well_control.R
\name{logbook_category_species_forbidden_well_control}
\alias{logbook_category_species_forbidden_well_control}
\title{Gives the inconsistencies between the weight categories and the species in the well}
\usage{
logbook_category_species_forbidden_well_control(
  dataframe1,
  output,
  species = c("SKJ"),
  weight_category = c("W-9")
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_category_species_forbidden_well_control () function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{species}{{\link[base]{character}} expected. Default values: c("SKJ"). Vector of species that must not have certain weight categories in well (weight_category)}

\item{weight_category}{{\link[base]{character}} expected. Default values: c("W-9"). Vector of weight category codes that must not have certain species in well (species)}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the logbook_category_species_forbidden_well_control function is to provide a table of data that contains an inconsistency between the weight categories and the species in the well
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  wellactivityspecies_id}}
\item{\code{  species_fao_code}}
\item{\code{  weightcategory_code}}
}
}
\examples{
#Sample well 1 and 2 are ok,
#Sample well 3 is associated with the species concerned (species) and the weight category
#              concerned (weight_category)
dataframe1 <- data.frame(wellactivityspecies_id = c("1", "2", "3"),
                         weightcategory_code = c("W-1", "W-9", "W-9"),
                         species_fao_code = c("SKJ", "ALB", "SKJ"))
logbook_category_species_forbidden_well_control(dataframe1, output = "report")
}
