% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbook_distribution_control.R
\name{logbook_distribution_control}
\alias{logbook_distribution_control}
\title{Gives the inconsistencies between the weights of small and big sample fish and the sum of the small and big weights in the associated well}
\usage{
logbook_distribution_control(
  dataframe1,
  dataframe2,
  dataframe3,
  dataframe4,
  output,
  species_category_small_big = c("ALB", "YFT", "BET", "SKJ"),
  species_category_unknown = c("SKJ"),
  weight_category_small = c("W-1"),
  weight_category_big = c("W-2"),
  weight_category_unknown = c("W-9")
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_distribution_control () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_distribution_control () function.}

\item{dataframe3}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_distribution_control () function.}

\item{dataframe4}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_distribution_control () function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{species_category_small_big}{{\link[base]{character}} expected. Default values: c("ALB", "YFT", "BET", "SKJ"). List of the inventory of species (FAO code) used to calculate weight category small and big in well}

\item{species_category_unknown}{{\link[base]{character}} expected. Default values: c("SKJ"). Vector of species categorized as small if weight category information is missing}

\item{weight_category_small}{{\link[base]{character}} expected. Default values: c("W-1"). Vector of small weight category codes}

\item{weight_category_big}{{\link[base]{character}} expected. Default values: c("W-2"). Vector of big weight category codes}

\item{weight_category_unknown}{{\link[base]{character}} expected. Default values: c("W-9"). Vector of unknown weight category codes}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the logbook_distribution_control  function is to provide a table of data that contains an inconsistency between the small and large sample weights and the sum of the small and big weights of the associated well
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  sample_id}}
\item{\code{  sample_well}}
\item{\code{  trip_id}}
\item{\code{  sample_smallsweight}}
\item{\code{  sample_bigsweight}}
}
\itemize{
Dataframe 2:
\item{\code{  well_id}}
\item{\code{  well_label}}
\item{\code{  trip_id}}
}
\itemize{
Dataframe 3:
\item{\code{  wellactivity_id}}
\item{\code{  well_id}}
}
\itemize{
Dataframe 4:
\item{\code{  wellactivityspecies_id}}
\item{\code{  wellactivity_id}}
\item{\code{  weightcategory_code}}
\item{\code{  species_fao_code}}
\item{\code{  wellactivityspecies_weight}}
}
}
\examples{
#Sample 1 and 2 are ok,
#Sample 3 has not small weight in well,
#Sample 4 has not bigs weight in sample,
#Sample 5 has different bigs weight,
#Sample 6 and 7 has different small weight
dataframe1 <- data.frame(sample_id = c("1", "2", "3", "4", "5", "6", "7"),
                         sample_well = c("well_1", "well_2", "well_3", "well_4", "well_5",
                                         "well_6","well_7"),
                         trip_id = c("1", "1", "1", "1", "1", "1", "1"),
                         sample_smallsweight = c(6, 25, 14, 0, NA, 10, 8),
                         sample_bigsweight = c(12, 0, 9, NA, 6, 0, 0))
dataframe2 <- data.frame(well_id = c("1", "2", "3", "4", "5", "6", "7"),
                         well_label = c("well_1", "well_2", "well_3", "well_4", "well_5", "well_6",
                                        "well_7"),
                         trip_id = c("1", "1", "1", "1", "1", "1", "1"))
dataframe3 <- data.frame(wellactivity_id = c("1", "2", "3", "4", "5", "6", "7", "8"),
                         well_id = c("1", "1", "2", "3", "4", "5", "6", "7"))
dataframe4 <- data.frame(wellactivityspecies_id = c("1", "2", "3", "4", "5", "6", "7", "8", "9",
                                                    "10", "11", "12", "13", "14"),
                         wellactivity_id = c("1", "1", "1", "2", "3", "4", "4", "5", "6", "7",
                                             "7", "7", "8", "8"),
                         weightcategory_code = c("W-1", "W-9", "W-2", "W-2", "W-1", "W-2", "W-9",
                                                 "W-2", "W-2", "W-1", "W-9", "W-9", "W-1", "W-1"),
                         species_fao_code = c("BET", "SKJ", "SKJ", "ALB", "SKJ", "BET", "BET",
                                              "ALB", "SKJ", "SKJ", "SKJ", "BET", "ALB", "JOS"),
                         wellactivityspecies_weight = c(4, 2, 7, 5, 25, 9, 14, 5, 17, 10, 5, 2, 7,
                                                        1))
logbook_distribution_control(dataframe1, dataframe2, dataframe3, dataframe4, output = "report")
}
