% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbook_fishing_context_control.R
\name{logbook_fishing_context_control}
\alias{logbook_fishing_context_control}
\title{Gives the inconsistencies between the school type and the association}
\usage{
logbook_fishing_context_control(
  dataframe1,
  dataframe2,
  output,
  school_type_object = c("1"),
  school_type_free = c("2"),
  school_type_unknown = c("0")
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_fishing_context_control () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_fishing_context_control () function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{school_type_object}{{\link[base]{character}} expected, default : c("1"). Vector of inventory of code of the object school.}

\item{school_type_free}{{\link[base]{character}} expected, default : c("2"). Vector of inventory of code of the free school.}

\item{school_type_unknown}{{\link[base]{character}} expected, default : c("0"). Vector of inventory of code of the unknown school.}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the logbook_fishing_context_control function is to provide a table of data that contains an inconsistency with school type and the association
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  activity_id}}
\item{\code{  schooltype_code}}
}
\itemize{
Dataframe 2:
\item{\code{  observedsystem_id}}
\item{\code{  activity_id}}
\item{\code{  schooltype_code}}
}
}
\examples{
#Activity 1, 2, 5, 6 and 7 are ok,
#Activity 3 has school type object,
#Activity 4 has no school type object
dataframe1 <- data.frame(activity_id = c("1", "2", "3", "4", "5", "6", "7"),
                         schooltype_code = c("2", "1", "2", "1", NA, "0", "2"))
dataframe2 <- data.frame(observedsystem_id = c("1", "2", "3", "4", "5"),
                         activity_id = c("1", "2", "2", "3", "4"),
                         schooltype_code = c("2", "2", "1", "1", "2"))
logbook_fishing_context_control(dataframe1, dataframe2, output = "report")
}
