% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/logbook_floating_object_buoy_id_redundancy_control.R
\name{logbook_floating_object_buoy_id_redundancy_control}
\alias{logbook_floating_object_buoy_id_redundancy_control}
\title{Identification of incoherent buoy operation on a same object}
\usage{
logbook_floating_object_buoy_id_redundancy_control(
  dataframe1,
  dataframe2,
  output,
  operation_code = c("2", "3")
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_floating_object_buoy_id_redundancy_control.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_floating_object_buoy_id_redundancy_control.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{operation_code}{{\link[base]{character}} expected. Default values: c("2","3"). Vector containing the operation code to detected to avoid incoherence.}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical".
}
\description{
The purpose of the logbook_floating_object_buoy_id_redundancy_control function is to provide a table of data that contains a incoherent buoy operation on a same object.
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  floatingobject_id}}
Dataframe 2:
\item{\code{  transmittingbuoy_id}}
\item{\code{  transmittingbuoyoperation_code}}
\item{\code{  floatingobject_id}}
\item{\code{  transmittingbuoy_code}}
}
}
\examples{
dataframe1 <- data.frame(floatingobject_id = c("1","2","3","4", "5"))
dataframe2 <- data.frame(transmittingbuoy_id = c("1","2","3","4", "5", "6"),
                         floatingobject_id = c("1","2","3","3","5","5"),
                         transmittingbuoyoperation_code = c("2","3","2","4", "2", "3"),
                         transmittingbuoy_code = c("1","2","3","3", "1", "1"))
transmittingbuoy_code = c("1", "2", "3", NA, "1"), logical = c(FALSE, FALSE, FALSE, FALSE, TRUE)), class = c("tbl_df", "tbl", "data.frame"), row.names = c(NA, -5L)))
logbook_floating_object_buoy_id_redundancy_control(dataframe1, dataframe2, output = "report")
}
