% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/logbook_floating_object_operation_buoy_control.R
\name{logbook_floating_object_operation_buoy_control}
\alias{logbook_floating_object_operation_buoy_control}
\title{Identification of incoherent between the floating object operation and buoy operation}
\usage{
logbook_floating_object_operation_buoy_control(
  dataframe1,
  dataframe2,
  output,
  operation_object_buoy = list(`1` = c("2", "4", "5"), `2` = c("4"), `4` = c("3", "4"),
    `5` = c("1", "3", "4"), `7` = c("3"), `8` = c("4"), `9` = c("4"), `6` = c("3"), `11`
    = c("1", "2", "3"), `12` = c("3", "1"))
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_floating_object_operation_buoy_control.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_floating_object_operation_buoy_control.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{operation_object_buoy}{{\link[base]{list}} expected. Default values: list("1" = c("2", "4", "5"), "2" = c("4"), "4" = c("3", "4"), "5" = c("1", "3", "4"), "7" = c("3"), "8" = c("4"), "9" = c("4"), "6" = c("3"), "11" = c("1", "2", "3"), "12" = c("3", "1")). List of the inventory of floating object operation (left side) prohibited with buoy operation used (right side)}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical".
}
\description{
The purpose of the logbook_floating_object_operation_buoy_control function is to provide a table of data that contains a incoherent with the floating object operation and buoy operation
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  transmittingbuoy_id}}
\item{\code{  transmittingbuoyoperation_code}}
\item{\code{  floatingobject_id}}
Dataframe 2:
\item{\code{  floatingobject_id}}
\item{\code{  objectoperation_code}}
}
}
\examples{
# Buoy 1 and 2 are ok,
# Buoy 3 has incompatible buoy operations and object operations
dataframe1 <- data.frame(transmittingbuoy_id = c("1","2","3"),
                         transmittingbuoyoperation_code = c("1", "4", "1"),
                         floatingobject_id = c("1","2","2"))
dataframe2 <- data.frame(floatingobject_id = c("1","2"),
                         objectoperation_code = c("1", "11"))
logbook_floating_object_operation_buoy_control(dataframe1, dataframe2, output = "report")
}
