% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/logbook_floating_object_part_DFAD_when_leaving_control.R
\name{logbook_floating_object_part_DFAD_when_leaving_control}
\alias{logbook_floating_object_part_DFAD_when_leaving_control}
\title{Identification of incoherent between the number of element DFAD when leaving and the object operation}
\usage{
logbook_floating_object_part_DFAD_when_leaving_control(
  dataframe1,
  dataframe2,
  output,
  object_operation = c("1")
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_floating_object_part_DFAD_when_leaving_control.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_floating_object_part_DFAD_when_leaving_control.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{object_operation}{{\link[base]{character}} expected. Default values: c("1"). Vector containing the codes for object operation with traditionally has a DFAD element when leaving}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical".
}
\description{
The purpose of the logbook_floating_object_part_DFAD_when_leaving_control function is to provide a table of data that contains a incoherent with the the number of element of part of objects DFAD for floating object when leaving and the object operation
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  floatingobject_id}}
\item{\code{  objectoperation_code}}
Dataframe 2:
\item{\code{  floatingobjectpart_id}}
\item{\code{  objectmaterial_code}}
\item{\code{  floatingobjectpart_whenleaving}}
\item{\code{  floatingobject_id}}
}
}
\examples{
#Floating object 1, 2 and 3 are ok,
#Floating object 4 has no elements 1-1 when leaving
#Floating object 5 has no elements 1-1
#Floating object 6 has no elements
dataframe1 <- data.frame(floatingobject_id = c("1", "2", "3", "4", "5", "6"),
                         objectoperation_code = c("2", "1", "1", "1", "1", "1"))
dataframe2 <- data.frame(floatingobjectpart_id = c("1", "2", "3", "4", "5", "6"),
                         objectmaterial_code = c("2-1", "1-1-1", "1-1", "1-1-1", "1-1-2", "2-1"),
                         floatingobjectpart_whenleaving = c("true", "true", "true", "false", NA,
                                                            "true"),
                         floatingobject_id = c("1", "2", "3", "4", "4", "5"))
logbook_floating_object_part_DFAD_when_leaving_control(dataframe1, dataframe2, output = "report")
}
