% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/logbook_floating_object_part_dimensions_when_leaving_control.R
\name{logbook_floating_object_part_dimensions_when_leaving_control}
\alias{logbook_floating_object_part_dimensions_when_leaving_control}
\title{Identification of incoherent between the number of dimensions element DFAD when leaving and the object operation}
\usage{
logbook_floating_object_part_dimensions_when_leaving_control(
  dataframe1,
  dataframe2,
  output,
  object_operation = c("1"),
  object_material_code = c("1-1-1-1-6", "1-1-1-1-7", "1-1-1-1-8", "1-1-1-1-9",
    "1-1-2-10", "1-1-2-11", "1-1-2-12", "1-1-2-13"),
  threshold = 2
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_floating_object_part_dimensions_when_leaving_control.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_floating_object_part_dimensions_when_leaving_control.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{object_operation}{{\link[base]{character}} expected. Default values: c("1"). Vector containing the codes for object operation with traditionally has a DFAD element when leaving}

\item{object_material_code}{{\link[base]{character}} expected. Default values: c("1-1-1-1-6", "1-1-1-1-7", "1-1-1-1-8", "1-1-1-1-9", "1-1-2-10", "1-1-2-11", "1-1-2-12", "1-1-2-13"). Vector containing the codes for object material with traditionally has a DFAD element when leaving}

\item{threshold}{{\link[base]{numeric}} expected. Default values: 2. Minimum number of object material requirements}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical".
}
\description{
The purpose of the logbook_floating_object_part_dimensions_when_leaving_control function is to provide a table of data that contains a incoherent with the the number of dimensions element of part of objects DFAD for floating object when leaving and the object operation
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  floatingobject_id}}
\item{\code{  objectoperation_code}}
Dataframe 2:
\item{\code{  floatingobjectpart_id}}
\item{\code{  objectmaterial_code}}
\item{\code{  floatingobjectpart_whenleaving}}
\item{\code{  floatingobject_id}}
}
}
\examples{
#Floating object 1, and 2 is ok,
#Floating object 3 does not have the required number (threshold) of elements (object_material_code)
#Floating object 4 has no elements object material
dataframe1 <- data.frame(floatingobject_id = c("1", "2", "3", "4"),
                         objectoperation_code = c("1", "8", "1", "1"))
dataframe2 <- data.frame(floatingobjectpart_id = c("1", "2", "3", "4"),
                         objectmaterial_code = c("1-1-1-1-6", "1-1-2-11", "1-1-1-1-7", "1-1"),
                         floatingobjectpart_whenleaving = c("3", "5", "1", "true"),
                         floatingobject_id = c("1", "1", "3", "3"))
logbook_floating_object_part_dimensions_when_leaving_control(dataframe1,
                                                             dataframe2,
                                                             output = "report")
}
