% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbook_floating_object_part_material_control.R
\name{logbook_floating_object_part_material_control}
\alias{logbook_floating_object_part_material_control}
\title{Identification of incoherent floating object material}
\usage{
logbook_floating_object_part_material_control(
  dataframe1,
  dataframe2,
  output,
  object_material = c("1-2")
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_floating_object_part_material_control.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_floating_object_part_material_control.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{object_material}{{\link[base]{character}} expected. Default values: c("1-2"). Vector containing the codes for materials that are not traditionally used on a floating object.}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical".
}
\description{
The purpose of the logbook_floating_object_part_material_control function is to provide a table of data that contains a incoherent floating object material. Indeed, certain floating objects are not used in check fishing.
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  floatingobject_id}}
Dataframe 2:
\item{\code{  floatingobjectpart_id}}
\item{\code{  objectmaterial_code}}
\item{\code{  floatingobject_id}}
}
}
\examples{
#Floating object 1 is ok,
#Floating object 2 has object material not used in the fishing.
dataframe1 <- data.frame(floatingobject_id = c("1", "2"))
dataframe2 <- data.frame(floatingobjectpart_id = c("1", "2"),
                         objectmaterial_code = c("1-1", "1-2"),
                         floatingobject_id = c("1", "2"))
logbook_floating_object_part_material_control(dataframe1, dataframe2, output = "report")
}
