% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/logbook_floating_object_part_when_leaving_control.R
\name{logbook_floating_object_part_when_leaving_control}
\alias{logbook_floating_object_part_when_leaving_control}
\title{Identification of incoherent between the number of part of objects for floating object when arriving and when leaving}
\usage{
logbook_floating_object_part_when_leaving_control(
  dataframe1,
  dataframe2,
  output,
  object_operation = c("8")
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_floating_object_part_when_leaving_control.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_floating_object_part_when_leaving_control.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{object_operation}{{\link[base]{character}} expected. Default values: c("8"). Vector containing the codes for object operation with traditionally more items at the end than at the beginning}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical".
}
\description{
The purpose of the logbook_floating_object_part_when_leaving_control function is to provide a table of data that contains a incoherent with the number of part of objects for floating object when arriving are  greater than  when leaving
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  floatingobject_id}}
\item{\code{  objectoperation_code}}
Dataframe 2:
\item{\code{  floatingobjectpart_id}}
\item{\code{  floatingobjectpart_whenarriving}}
\item{\code{  floatingobjectpart_whenleaving}}
\item{\code{  floatingobject_id}}
}
}
\examples{
#Floating object 1, 2 and 3 is ok,
#Floating object 4 has no elements that have been added
#Floating object 5 does not include all of these elements that have been preserved
dataframe1 <- data.frame(floatingobject_id = c("1", "2", "3", "4", "5"),
                         objectoperation_code = c("1", "8", "8", "8", "8"))
dataframe2 <- data.frame(floatingobjectpart_id = c("1", "2", "3", "4", "5", "6", "7", "8"),
                         floatingobjectpart_whenarriving = c("true", NA, "false", "true", NA,
                                                             "true", "true", NA),
                         floatingobjectpart_whenleaving = c(NA, "true", "true", "true", "9",
                                                            "true", "false", "true"),
                         floatingobject_id = c("1", "2", "2", "3", "3", "4", "5", "5"))
logbook_floating_object_part_when_leaving_control(dataframe1, dataframe2, output = "report")
}
