% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbook_landing_control.R
\name{logbook_landing_control}
\alias{logbook_landing_control}
\title{Gives the inconsistencies between the total landed weight for canneries and local market in the trip and vessel capacity link to trip}
\usage{
logbook_landing_control(dataframe1, output)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_landing_control () function.}

\item{output}{{\link[base]{character}} expected.Kind of expected output. You can choose between "message", "report" or "logical".}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the logbook_landing_control function is to provide a table of data that contains an inconsistency with the total landed weight greater than vessel capacity for the trip
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  trip_id}}
\item{\code{  trip_landingtotalweight}}
\item{\code{  trip_localmarkettotalweight}}
\item{\code{  vessel_capacity}}
}
}
\examples{
#Trip 1 and 4 are ok,
#Trip 2 has a total weight greater than the vessel's capacity,
#Trip 3 has its vessel's capacity missing
dataframe1 <- data.frame(trip_id = c("1", "2", "3", "4"),
                         trip_landingtotalweight = c(10, 15, 2, NA),
                         trip_localmarkettotalweight = c(2, 1, 7, NA),
                         vessel_capacity = c(20, 10, NA, 13))
logbook_landing_control(dataframe1, output = "report")
}
