% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbook_ldlf_control.R
\name{logbook_ldlf_control}
\alias{logbook_ldlf_control}
\title{Gives the inconsistencies between the sample measurement types and species or weight values}
\usage{
logbook_ldlf_control(
  dataframe1,
  dataframe2,
  output,
  species = c("SKJ", "LTA", "FRI"),
  size_measure_type_species = c("PD1"),
  size_measure_type_big = c("PD1"),
  size_measure_type_small = c("FL")
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_ldlf_control () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_ldlf_control () function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{species}{{\link[base]{character}} expected. Default values:  c("SKJ", "LTA", "FRI"). Vector of the species not to be associated with a type of measure.}

\item{size_measure_type_species}{{\link[base]{character}} expected. Default values:  c("PD1"). Vector of type of measure not to be associated with species}

\item{size_measure_type_big}{{\link[base]{character}} expected. Default values:  c("PD1"). Type of measure that must have a total weight or a big fish weight}

\item{size_measure_type_small}{{\link[base]{character}} expected. Default values: c("FL"). Type of measure that must have a total weight or a small fish weight}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the logbook_ldlf_control  function is to provide a table of data that contains an inconsistency between the sample measurement types and species or weight values
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  samplespecies_id}}
\item{\code{  species_fao_code}}
\item{\code{  sizemeasuretype_code}}
\item{\code{  sample_id}}
}
\itemize{
Dataframe 2:
\item{\code{  sample_id}}
\item{\code{  sample_smallsweight}}
\item{\code{  sample_bigsweight}}
\item{\code{  sample_totalweight}}
}
}
\examples{
#Sample species 1, 3 and 4 are ok,
#Sample species 2 has a species that is not compatible with the type of measure,
#Sample species 5 has no smalls weight nor total weight,
#Sample species 6 has no bigs weight nor total weight
dataframe1 <- data.frame(samplespecies_id = c("1", "2", "3", "4", "5", "6"),
                         species_fao_code = c("SKJ", "SKJ", "JOS", "SKJ", "SKJ", "JOS"),
                         sizemeasuretype_code = c("FL", "PD1", "PD1", "FL", "FL", "PD1"),
                         sample_id = c("1", "1", "1", "2", "3", "4"))
dataframe2 <- data.frame(sample_id = c("1", "2", "3", "4"),
                         sample_smallsweight = c(5, 0, 0, 7),
                         sample_bigsweight = c(12, NA, 6, 0),
                         sample_totalweight = c(0, 23, NA, NA))
logbook_ldlf_control(dataframe1, dataframe2, output = "report")
}
