% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbook_length_class_control.R
\name{logbook_length_class_control}
\alias{logbook_length_class_control}
\title{Gives the inconsistencies between size class of the samples depending on the species and measurement type and the valid threshold}
\usage{
logbook_length_class_control(
  dataframe1,
  output,
  size_measure_type = c("FL", "PD1"),
  threshold = c(10, 80)
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_length_class_control () function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{size_measure_type}{{\link[base]{character}} expected. Default values: c("FL", "PD1"). Vector of the size measure type controlled.}

\item{threshold}{{\link[base]{numeric}} expected. Default values: 10 and 80. Vector containing the lower and upper acceptable threshold for the size measured}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the logbook_length_class_control function is to provide a table of data that contains an inconsistency between the size class of the samples depending on the species and measurement type and the valid threshold (Default : 80)
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  samplespeciesmeasure_id}}
\item{\code{  sizemeasuretype_code}}
\item{\code{  samplespeciesmeasure_sizeclass}}
}
}
\examples{
#Activity 1 and 2 are ok,
#Activity 3 has size class is less than threshold,
#Activity 4 has size class is greater than threshold,
#Activity 5 has size class is missing
dataframe1 <- data.frame(samplespeciesmeasure_id = c("1", "2", "3", "4", "5"),
                         sizemeasuretype_code = c("FL", "PD1", "PD1", "FL", "FL"),
                         samplespeciesmeasure_sizeclass = c(25, 75, 5, 83, NA))
logbook_length_class_control(dataframe1, output = "report")
}
