% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbook_measure_control.R
\name{logbook_measure_control}
\alias{logbook_measure_control}
\title{Gives the inconsistencies between the total number of individuals measured per sample and the sum of individuals per sample, species and size class}
\usage{
logbook_measure_control(dataframe1, dataframe2, output)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_measure_control () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_measure_control () function.}

\item{output}{{\link[base]{character}} expected.Kind of expected output. You can choose between "message", "report" or "logical".}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the logbook_measure_control function is to provide a table of data that contains an inconsistency between the total number of individuals measured per sample and the sum of individuals per sample, species and size class
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  samplespecies_id}}
\item{\code{  samplespecies_measuredcount}}
\item{\code{  sample_id}}
}
\itemize{
Dataframe 2:
\item{\code{  samplespeciesmeasure_id}}
\item{\code{  samplespeciesmeasure_count}}
\item{\code{  samplespecies_id}}
}
}
\examples{
#Sample 1 and 2 are ok,
#Sample 3 has different count,
#Sample 4 has sample species measure count is missing in dataframe2,
#Sample 5 has sample species measured count is missing in dataframe1
dataframe1 <- data.frame(samplespecies_id = c("1", "2", "3", "4", "5", "6"),
                         samplespecies_measuredcount = c(4, 6, 15, 6, 7, NA),
                         sample_id = c("1", "1", "2", "3", "4", "5"))
dataframe2 <- data.frame(samplespeciesmeasure_id = c("1", "2", "3", "4", "5", "6"),
                         samplespeciesmeasure_count = c(10, 10, 5, 3, 2, 8),
                         samplespecies_id = c("1", "3", "3", "4", "4", "6"))
logbook_measure_control(dataframe1, dataframe2, output = "report")
}
