% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/logbook_observed_system_floatting_object_pair_control.R
\name{logbook_observed_system_floatting_object_pair_control}
\alias{logbook_observed_system_floatting_object_pair_control}
\title{Identification of incoherent between the presence of floatting object and observed system}
\usage{
logbook_observed_system_floatting_object_pair_control(
  dataframe1,
  dataframe2,
  dataframe3,
  output,
  vessel_activity = c("6"),
  observed_system = c("20")
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_observed_system_floatting_object_pair_control.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_observed_system_floatting_object_pair_control.}

\item{dataframe3}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_observed_system_floatting_object_pair_control.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{vessel_activity}{{\link[base]{character}} expected. Default values: c("6"). Vector containing the code for the vessel activity, which, if it contains a floating object, must also have an observer system (code observed_system) and vice versa.}

\item{observed_system}{{\link[base]{character}} expected. Default values: c("20"). Vector containing the code of the observed system that must be linked to the vessel activity}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical".
}
\description{
The purpose of the logbook_observed_system_floatting_object_pair_control function is to provide a table of data that contains a incoherent between the presence of floatting object and observed system for certain vessel activity.
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  activity_id}}
\item{\code{  vesselactivity_code}}
}
\itemize{
Dataframe 2:
\item{\code{  floatingobject_id}}
\item{\code{  activity_id}}
}
\itemize{
Dataframe 3:
\item{\code{  observedsystem_id}}
\item{\code{  observedsystem_code}}
\item{\code{  activity_id}}
}
}
\examples{
#Activity 1, 2, 3 and 4 are ok,
#Activity 5 has no observed system 20
#Activity 6 has observed system missing
#Activity 7 has floating object missing
#Activity 8 has vessel activity missing
dataframe1 <- data.frame(activity_id = c("1", "2", "3", "4", "5", "6", "7", "8"),
                         vesselactivity_code = c("6", "6", "1", "2", "6", "6", "6", NA))
dataframe2 <- data.frame(floatingobject_id = c("1", "2", "3", "4", "5"),
                         activity_id = c("1", "4", "5", "6", "8"))
dataframe3 <- data.frame(observedsystem_id = c("1", "2", "3", "4", "5", "6"),
                         observedsystem_code = c("20", "2", "20", "2", "20", "20"),
                         activity_id = c("1", "2", "3", "5", "7", "8"))
logbook_observed_system_floatting_object_pair_control(dataframe1,
                                                      dataframe2,
                                                      dataframe3,
                                                      output = "report")
}
