% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbook_operationobjet_observedsystem_control.R
\name{logbook_operationobjet_observedsystem_control}
\alias{logbook_operationobjet_observedsystem_control}
\title{Identification of incoherent operation objet and observed system}
\usage{
logbook_operationobjet_observedsystem_control(
  dataframe1,
  dataframe2,
  output,
  operation_objet = c("1"),
  observed_system = c("20")
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_operationobjet_observedsystem_control.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_operationobjet_observedsystem_control.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{operation_objet}{{\link[base]{character}} expected. Default values: c("1"). Vector containing the code for the operation on the object that must not have the observed system.}

\item{observed_system}{{\link[base]{character}} expected. Default values: c("20"). Vector containing the code of the observed system that must not be linked to the object operation.}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical".
}
\description{
The purpose of the logbook_operationobjet_observedsystem_control function is to provide a table of data that contains a incoherent operation objet and the presence of observed system.
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  floatingobject_id}}
\item{\code{  objectoperation_code}}
\item{\code{  activity_id}}
Dataframe 2:
\item{\code{  observedsystem_id}}
\item{\code{  observedsystem_code}}
\item{\code{  activity_id}}
}
}
\examples{
#Floating object 1, 2 and 3 are ok,
#Floating object 4 has a conflict between the operation on the object and the observed system
#Floating object 5 has operation on the object missing
dataframe1 <- data.frame(floatingobject_id = c("1", "2", "3", "4", "5"),
                         objectoperation_code = c("1", "1", "2", "1", NA),
                         activity_id = c("1", "2", "3", "3", "4"))
dataframe2 <- data.frame(observedsystem_id = c("1", "2", "3"),
                         observedsystem_code = c("1", "2", "20"),
                         activity_id = c("1", "1", "3"))
logbook_operationobjet_observedsystem_control(dataframe1, dataframe2, output = "report")
}
