% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbook_sampleactivity_well_sampled_control.R
\name{logbook_sampleactivity_well_sampled_control}
\alias{logbook_sampleactivity_well_sampled_control}
\title{Gives the inconsistencies between the sampling activity and well sampled plan trip}
\usage{
logbook_sampleactivity_well_sampled_control(dataframe1, dataframe2, output)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_sampleactivity_well_sampled_control () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_sampleactivity_well_sampled_control () function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the logbook_sampleactivity_well_sampled_control  function is to provide a table of data that contains an inconsistency between the sampling activity and well sampled plan trip
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  wellactivity_id}}
\item{\code{  well_label}}
\item{\code{  activity_id}}
\item{\code{  well_id}}
}
\itemize{
Dataframe 2:
\item{\code{  sampleactivity_id}}
\item{\code{  sample_well}}
\item{\code{  activity_id}}
}
}
\examples{
#Well activity 1, 2 and 3 are ok,
#Well activity 4 is not linked to a sample, even though the well is sampled
dataframe1 <- data.frame(wellactivity_id = c("1", "2", "3", "4"),
                         well_label = c("well_1", "well_2", "well_3", "well_3"),
                         activity_id = c("1", "2", "3", "4"),
                         well_id = c("1", "2", "3", "3"))
dataframe2 <- data.frame(sampleactivity_id = c("1", "2"),
                         sample_well = c("well_1", "well_3"),
                         activity_id = c("1", "3"))
logbook_sampleactivity_well_sampled_control(dataframe1, dataframe2, output = "report")
}
