% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbook_sea_time_control.R
\name{logbook_sea_time_control}
\alias{logbook_sea_time_control}
\title{Gives the inconsistencies between the sum of the sea times indicated for the route and the one indicated for the trip}
\usage{
logbook_sea_time_control(dataframe1, dataframe2, output)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_sea_time_control () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_sea_time_control () function.}

\item{output}{{\link[base]{character}} expected.Kind of expected output. You can choose between "message", "report" or "logical".}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the logbook_sea_time_control function is to provide a table of data that contains an inconsistency between the sum of the sea times indicated for the route and the one indicated for the trip
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  trip_id}}
\item{\code{  trip_seatime}}
}
\itemize{
Dataframe 2:
\item{\code{  route_id}}
\item{\code{  trip_id}}
\item{\code{  route_seatime}}
}
}
\examples{
#Trip 1 is ok,
#Trip 2 has different sea times,
#Trip 3 has no trip_seatime,
#Trip 4 has no route_seatime,
#Trip 5 has trip_seatime = 0
dataframe1 <- data.frame(trip_id = c("1", "2", "3", "4", "5"),
                         trip_seatime = c(10, 15, NA, 8, 0))
dataframe2 <- data.frame(route_id = c("1", "2", "3", "4", "5", "6"),
                         trip_id = c("1", "1", "2", "2", "3", "5"),
                         route_seatime = c(4, 6, 10, 6, 5, 0))
logbook_sea_time_control(dataframe1, dataframe2, output = "report")
}
