% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbook_super_sample_number_control.R
\name{logbook_super_sample_number_control}
\alias{logbook_super_sample_number_control}
\title{Gives the inconsistencies between the sample and the subsample number}
\usage{
logbook_super_sample_number_control(dataframe1, dataframe2, output)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_super_sample_number_control() function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_super_sample_number_control() function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the logbook_super_sample_number_control function is to provide a table of data that contains an inconsistency between the sample and the subsample number
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  sample_id}}
\item{\code{  sample_supersample}}
}
\itemize{
Dataframe 2:
\item{\code{  samplespecies_id}}
\item{\code{  samplespecies_subsamplenumber}}
\item{\code{  sample_id}}
}
}
\examples{
#Sample 1 and 2 are ok,
#Sample 3 is not a super sample but the numbering starts at 1 instead of 0,
#Sample 4 is a super sample but the same sub-sample number appears twice,
#Sample 5 has no sample species,
#Sample 6 is a super sample but the numbering starts at 0 instead of 1,
#Sample 7 is a super sample but has only one sub-sample species
#Sample 8 is a super sample but has discontinuous numbering of sub-samples
dataframe1 <- data.frame(sample_id = c("1", "2", "3", "4", "5", "6", "7", "8"),
                         sample_supersample = c(FALSE, TRUE, FALSE, TRUE, FALSE, TRUE, TRUE, TRUE))
dataframe2 <- data.frame(samplespecies_id = c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10",
                                              "11", "12", "13"),
                         samplespecies_subsamplenumber = c(0, 0, 1, 1, 2, 1, 1, 1, 0, 1, 1, 1, 3),
                         sample_id = c("1", "1", "2", "2", "2", "3", "4", "4", "6", "6", "7", "8",
                                       "8"))
logbook_super_sample_number_control(dataframe1, dataframe2, output = "report")
}
