% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbook_trip_activity_control.R
\name{logbook_trip_activity_control}
\alias{logbook_trip_activity_control}
\title{Gives the inconsistencies between the trip and the associated activities in terms of presence}
\usage{
logbook_trip_activity_control(dataframe1, dataframe2, output)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_trip_activity_control () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_trip_activity_control () function.}

\item{output}{{\link[base]{character}} expected.Kind of expected output. You can choose between "message", "report" or "logical".}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the logbook_trip_activity_control function is to provide a table of data that contains an inconsistency between the trip and the presence or not of activity associated with this trip
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  trip_id}}
}
\itemize{
Dataframe 2:
\item{\code{  trip_id}}
\item{\code{  activity_id}}
}
}
\examples{
#Trip 1 is ok,
#Trip 2 is not linked to an activity
dataframe1 <- data.frame(trip_id = c("1", "2"))
dataframe2 <- data.frame(trip_id = c("1"),
                         activity_id = c("1"))
logbook_trip_activity_control(dataframe1, dataframe2, output = "report")
}
