% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/logbook_vessel_activity_object_operation_control.R
\name{logbook_vessel_activity_object_operation_control}
\alias{logbook_vessel_activity_object_operation_control}
\title{Identification of incoherent vessel activity and object operation}
\usage{
logbook_vessel_activity_object_operation_control(
  dataframe1,
  dataframe2,
  output,
  vessel_activity = c("6"),
  operation_objet = c("1")
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_vessel_activity_object_operation_control.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_vessel_activity_object_operation_control.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{vessel_activity}{{\link[base]{character}} expected. Default values: c("6"). Vector containing the vessel activity code that must not have a floating object with this object operation (operation_objet).}

\item{operation_objet}{{\link[base]{character}} expected. Default values: c("1"). Vector containing the code for the operation on the object that must not have this vessel activity (vessel_activity).}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical".
}
\description{
The purpose of the logbook_vessel_activity_object_operation_control function is to provide a table of data that contains a incoherent vessel activity and object operation for floating object.
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  activity_id}}
\item{\code{  vesselactivity_code}}
Dataframe 2:
\item{\code{  floatingobject_id}}
\item{\code{  objectoperation_code}}
\item{\code{  activity_id}}
}
}
\examples{
#Activity 1 and 2 are ok,
#Activity 3 has conflit between vessel activity and operation objet
#Activity 4 has vessel activity missing
#Activity 5 has operation objet missing
dataframe1 <- data.frame(activity_id = c("1", "2", "3", "4", "5"),
                         vesselactivity_code = c("6", "1", "6", NA, "6"))
dataframe2 <- data.frame(floatingobject_id = c("1", "2", "3", "4"),
                         objectoperation_code = c("2", "1", "1", NA),
                         activity_id = c("1", "2", "3", "5"))
logbook_vessel_activity_object_operation_control(dataframe1, dataframe2, output = "report")
}
