% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbook_weigh_catch_well_control.R
\name{logbook_weigh_catch_well_control}
\alias{logbook_weigh_catch_well_control}
\title{Gives the inconsistencies between the catch weight and weight in well}
\usage{
logbook_weigh_catch_well_control(
  dataframe1,
  dataframe2,
  output,
  species = c("YFT", "SKJ", "BET", "ALB"),
  species_fate = c("6", "15"),
  species_grouped_well = ("XXX*")
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_weigh_catch_well_control () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_weigh_catch_well_control () function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{species}{{\link[base]{character}} expected. Default values: c("YFT", "SKJ", "BET", "ALB"). list of of species (FAO code) listed separately in the well plan.}

\item{species_fate}{{\link[base]{character}} expected. Default values: c("6", "15"). Vector of inventory of fate retained on board.}

\item{species_grouped_well}{{\link[base]{character}} expected. Default values: "XXX*". Code for accessory species grouped together within the well plan}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the logbook_weigh_catch_well_control function is to provide a table of data that contains an inconsistency between the catch weight and weight in well
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  catch_id}}
\item{\code{  catch_weight}}
\item{\code{  speciesfate_code}}
\item{\code{  species_fao_code}}
\item{\code{  activity_id}}
}
\itemize{
Dataframe 2:
\item{\code{  wellactivityspecies_id}}
\item{\code{  wellactivityspecies_weight}}
\item{\code{  species_fao_code}}
\item{\code{  activity_id}}
}
}
\examples{
#Activity 1 are ok,
#Activity 2 has different weight,
#Activity 3 has well missing,
#Activity 4 has catch missing,
dataframe1 <- data.frame(catch_id = c("1", "2", "3", "4", "5", "6"),
                         catch_weight = c(3, 2, 4, 1, 15, 6),
                         speciesfate_code = c("6", "6", "6", "11", "6", "6"),
                         species_fao_code = c("YFT", "SKJ", "FRI", "JOS", "ALB", "FRI"),
                         activity_id = c("1", "1", "1", "1", "2", "3"))
dataframe2 <- data.frame(wellactivityspecies_id = c("1", "2", "3", "4", "5"),
                         wellactivityspecies_weight = c(3, 2, 4, 13, 4),
                         species_fao_code = c("YFT", "SKJ", "XXX*", "ALB", "BET"),
                         activity_id = c("1", "1", "1", "2", "4"))
logbook_weigh_catch_well_control(dataframe1, dataframe2, output = "report")
}
