% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbook_weight_control.R
\name{logbook_weight_control}
\alias{logbook_weight_control}
\title{Gives the inconsistencies between the sum of the weight indicated for catches and the one indicated for the activity}
\usage{
logbook_weight_control(dataframe1, dataframe2, output, epsilon = 0.01)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_weight_control () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_weight_control () function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{epsilon}{{\link[base]{numeric}} expected, default : 0.01. Gives the threshold at which the difference is considered too large.}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the logbook_weight_control function is to provide a table of data that contains an inconsistency between the sum of the weight indicated for the catch and the one indicated for the activity
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  activity_id}}
\item{\code{  activity_weight}}
}
\itemize{
Dataframe 2:
\item{\code{  catch_id}}
\item{\code{  catch_weight}}
\item{\code{  activity_id}}
}
}
\examples{
#Activity 1, 2 and 4 are ok,
#Activity 3 has different weight
dataframe1 <- data.frame(activity_id = c("1", "2", "3", "4"),
                         activity_weight = c(10, 15.01, 6, NA))
dataframe2 <- data.frame(catch_id = c("1", "2", "3", "4", "5"),
                         catch_weight = c(10, 10, 5, 3, 2),
                         activity_id = c("1", "2", "2", "3", "3"))
logbook_weight_control(dataframe1, dataframe2, output = "report")
}
