% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbook_weighting_sample_control.R
\name{logbook_weighting_sample_control}
\alias{logbook_weighting_sample_control}
\title{Gives the inconsistencies between the sample weighting and weight in well}
\usage{
logbook_weighting_sample_control(
  dataframe1,
  dataframe2,
  output,
  species = c("YFT", "SKJ", "BET", "ALB", "LTA", "FRI", "TUN", "KAW", "LOT"),
  epsilon = 0.01
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_weighting_sample_control () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_weighting_sample_control () function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{species}{{\link[base]{character}} expected. Default values: c("YFT", "SKJ", "BET", "ALB", "LTA", "FRI", "TUN", "KAW", "LOT"). list of the inventory of species (FAO code) used to compare to sample weighting.}

\item{epsilon}{{\link[base]{numeric}} expected, default : 0.01. Gives the threshold at which the difference is considered too large.}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the logbook_weighting_sample_control function is to provide a table of data that contains an inconsistency between the sample weighting and weight in well
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  sampleactivity_id}}
\item{\code{  sampleactivity_weightedweight}}
\item{\code{  sample_well}}
\item{\code{  activity_id}}
\item{\code{  trip_id}}
}
\itemize{
Dataframe 2:
\item{\code{  wellactivityspecies_id}}
\item{\code{  wellactivityspecies_weight}}
\item{\code{  species_fao_code}}
\item{\code{  well_label}}
\item{\code{  activity_id}}
\item{\code{  trip_id}}
}
}
\examples{
#Sample activity 1, 2, 3, 4, 5, 6 and 7 are ok,
#Sample Activity 8 has weights missing from the well,
#Sample Activity 9 does not take into account all species
#Sample Activity 10 does not take into account all well
#Sample Activity 11 has weighted weight missing
dataframe1 <- data.frame(sampleactivity_id = c("1", "2", "3", "4", "5", "6", "7", "8", "9",
                                               "10", "11"),
                         sampleactivity_weightedweight = c(3, 6, 20, 6.25, 13.75, 7.6, 2, 2,
                                                           26, 13, NA),
                         sample_well = c("well_1", "well_2", "well_1", "well_1", "well_2",
                                         "well_1", "well_2", "well_1", "well_1", "well_1",
                                         "well_1"),
                         activity_id = c("1", "1", "2", "3", "3", "4", "5", "6", "7", "8", "9"),
                         trip_id = c("1", "1", "2", "3", "3", "4", "4", "5", "6", "7", "8"))
dataframe2 <- data.frame(wellactivityspecies_id = c("1", "2", "3", "4", "5", "6", "7", "8", "9",
                                                    "10", "11", "12", "13", "14", "15", "16", "17"),
                         wellactivityspecies_weight = c(3, 2, 4, 1, 15, 5, 5, 11, 4, 4, 6, 9,
                                                        2, 26, 2, 13, 6),
                         species_fao_code = c("YFT", "SKJ", "ALB", "JOS", "YFT", "FRI", "YFT",
                                              "SKJ", "BET", "YFT", "YFT", "ALB", "YFT", "ALB",
                                              "YFT", "BET", "ALB"),
                         well_label = c("well_1", "well_2", "well_2", "well_2", "well_1", "well_2",
                                        "well_1", "well_2", "well_3", "well_1", "well_2", "well_3",
                                        "well_2", "well_1", "well_1", "well_1", "well_2"),
                         activity_id = c("1", "1", "1", "1", "2", "2", "3", "3", "3", "4", "4", "4",
                                         "5", "7", "7", "8", "8"),
                         trip_id = c("1", "1", "1", "1", "2", "2", "3", "3", "3", "4", "4", "4",
                                     "4", "6", "6", "7", "7"))
logbook_weighting_sample_control(dataframe1, dataframe2, output = "report")
}
