% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbook_well_number_control.R
\name{logbook_well_number_control}
\alias{logbook_well_number_control}
\title{Gives the inconsistencies between sample well number and associated trip well numbers}
\usage{
logbook_well_number_control(
  dataframe1,
  dataframe2,
  dataframe3,
  output,
  vessel_type = c("5", "6", "10")
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_well_number_control () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_well_number_control () function.}

\item{dataframe3}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_well_number_control () function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{vessel_type}{{\link[base]{character}} expected. Default values: c("5", "6", "10"). Vector of codes for vessel types with a well plan.}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the logbook_well_number_control  function is to provide a table of data that contains an inconsistency between sample well number and associated trip well numbers
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  sample_id}}
\item{\code{  sample_well}}
\item{\code{  trip_id}}
}
\itemize{
Dataframe 2:
\item{\code{  well_id}}
\item{\code{  trip_id}}
\item{\code{  well_label}}
}
\itemize{
Dataframe 3:
\item{\code{  trip_id}}
\item{\code{  vesseltype_code}}
}
}
\examples{
#Sample 1, 2 and 3 are ok,
#Sample 4 has different well,
#Sample 5 has no well
dataframe1 <- data.frame(sample_id = c("1", "2", "3", "4", "5"),
                         sample_well = c("well_1", "well_2", "well_2", "well_2", NA),
                         trip_id = c("1", "1", "2", "3", "3"))
dataframe2 <- data.frame(well_id = c("1", "2", "3"),
                         trip_id = c("1", "1", "3"),
                         well_label = c("well_1", "well_2", "well_1"))
dataframe3 <- data.frame(trip_id = c("1", "2", "3"),
                         vesseltype_code = c("6", "1", "6"))
logbook_well_number_control(dataframe1, dataframe2, dataframe3, output = "report")
}
