% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marine_area_overlay_control.R
\name{marine_area_overlay_control}
\alias{marine_area_overlay_control}
\title{Marine area overlay control function}
\usage{
marine_area_overlay_control(
  data_longitude,
  data_latitude,
  overlay_expected,
  area_file_path,
  output
)
}
\arguments{
\item{data_longitude}{{\link[base]{character}} expected. Vector of longitude values.}

\item{data_latitude}{{\link[base]{character}} expected. Vector of latitude values.}

\item{overlay_expected}{{\link[base]{character}} expected. Type of overlay output. You can choose between "fao_area", "eez_area" or "ices_area".}

\item{area_file_path}{{\link[base]{character}} expected. File path of the area shape. You can provide a .shp or a .RData file.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}
}
\description{
Control processes associated to the function {\link[furdeb]{marine_area_overlay}} (consistent spatial marine area overlay for points, grids and polygons).
}
\examples{
\dontrun{
marine_area_overlay_control(data_longitude = as.character(c(1.09, 8.37, 19.35, -4.04, -4.54)),
                            data_latitude = as.character(c(2.47, -3.19, 24.55, 5.22, 49.73)),
                            overlay_expected = "ices_area",
                            area_file_path = "ICES_Statistical_Rectangles_Eco.RData",
                            output = "report")
}
}
