% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_type_control.R
\name{measure_type_control}
\alias{measure_type_control}
\title{Measure type control}
\usage{
measure_type_control(
  data_connection,
  start_year,
  end_year,
  program,
  ocean,
  country_code,
  path_file = NULL
)
}
\arguments{
\item{data_connection}{{\link[base]{list}} expected. Either output of the function {\link[furdeb]{postgresql_dbconnection}}, which must be done before using the measure_type_control.}

\item{start_year}{{\link[base]{integer}} expected. Starting year for the control.}

\item{end_year}{{\link[base]{integer}} expected. Ending year for the control.}

\item{program}{{\link[base]{character}} expected. Programs to be controlled. Example of the format for a program topiaid: "fr.ird.referential.ps.common.Program#1239832686262#0.31033946454061234".}

\item{ocean}{{\link[base]{character}} expected. Ocean to be controlled. Examples: 'Indian', 'Atlantic'...etc.}

\item{country_code}{{\link[base]{character}} expected. Countries on which control will be made. Examples: 'FRA', 'MUS'...etc.}

\item{path_file}{{\link[base]{character}} expected.  By default NULL. Path to save the final xlsx.}
}
\value{
The function return a xlsx table.
}
\description{
This function allows to check the coherence of a sampled species measure type according to its default measure type.
}
\note{
Version 1.0
}
\author{
Esther Mollier, Philippe S. Sabarros
}
