% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_type_correction.R
\name{measure_type_correction}
\alias{measure_type_correction}
\title{Measure type correction}
\usage{
measure_type_correction(
  data_connection,
  start_year,
  end_year,
  program,
  ocean,
  country_code,
  species_group,
  species,
  sizemeasuretype_to_replace,
  sizemeasuretype_new,
  corrector,
  action,
  path_file = NULL
)
}
\arguments{
\item{data_connection}{{\link[base]{list}} expected. Either output of the function {\link[furdeb]{postgresql_dbconnection}}, which must be done before using the measure_type_control.}

\item{start_year}{{\link[base]{integer}} expected. Starting year for the correction.}

\item{end_year}{{\link[base]{integer}} expected. Ending year for the correction.}

\item{program}{{\link[base]{character}} expected. Programs to be controlled. Example of the format for a program topiaid: "fr.ird.referential.ps.common.Program#1239832686262#0.31033946454061234".}

\item{ocean}{{\link[base]{character}} expected. Ocean to be controlled. Examples: 'Indian', 'Atlantic'...etc.}

\item{country_code}{{\link[base]{character}} expected. Countries on wich control will be made. Examples: 'FRA', 'MUS'...etc.}

\item{species_group}{{\link[base]{character}} expected. Specie group to correct.}

\item{species}{{\link[base]{character}} expected. Species code to correct. Examples: 'YFT', 'FAL'...etc.}

\item{sizemeasuretype_to_replace}{{\link[base]{character}} expected. Measure type of the sample we want to correct. Examples: 'FL', 'DW'...etc.}

\item{sizemeasuretype_new}{{\link[base]{character}} expected. Measure type to replace the incorrect one. Examples: 'FL', 'DW'...etc.}

\item{corrector}{{\link[base]{character}} expected. First letter of the corrector's first name and last name. Examples: 'JMartin' for Jeanne Martin.}

\item{action}{{\link[base]{character}} expected. Type of action required when update queries are launched. COMMIT is used to definitively validate modifications and make them permanent in the database. ROLLBACK is used to undo changes made.}

\item{path_file}{{\link[base]{character}} expected. By default NULL. Path to save the final xlsx.}
}
\value{
The function returns a xlsx table.
}
\description{
This function aims to correct directly in the data base samples with incorrect measure type.
}
\note{
Version 1.0
}
\author{
Chloé Tellier, Esther Mollier, Philippe S. Sabarros
}
