% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_catch_ocean_control.R
\name{obs_catch_ocean_control}
\alias{obs_catch_ocean_control}
\title{Catch ocean control}
\usage{
obs_catch_ocean_control(catch)
}
\arguments{
\item{catch}{{\link[base]{data.frame}} expected. All catches during the time range selected.}
}
\value{
The function returns one {\link[base]{data.frame}}.
}
\description{
Checks the coherence of the species caught according to their distribution area.
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe catch:
\item{\code{  catch_id}}
\item{\code{  species_faocode}}
\item{\code{  ocean_label}}
\item{\code{  speciesocean_label}}
}
}
\note{
Version 1.0
}
\author{
Chloé Tellier, Esther Mollier, Philippe S. Sabarros
}
\examples{
#Catch 1 is ok: a blue marlin found in the Indian ocean.
#Catch 2 is ok: a black marlin found in the Indian ocean.
#Catch 3 is not ok: a black marlin found in the Atlantic ocean.
catch <- data.frame(catch_id = c("1", "2", "3"),
                    species_faocode = c("BUM", "BLM", "BLM"),
                    ocean_label = c("Indian", "Indian", "Atlantic"),
                    speciesocean_label = c("Indian", "Indian", NA))
obs_catch_ocean_control(catch)
}
