% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/obs_catch_or_sample_fate_by_species_group_control.R
\name{obs_catch_or_sample_fate_by_species_group_control}
\alias{obs_catch_or_sample_fate_by_species_group_control}
\title{Catch or sample fate by species group control}
\usage{
obs_catch_or_sample_fate_by_species_group_control(data)
}
\arguments{
\item{data}{{\link[base]{data.frame}} expected. All catches or samples during the time range selected.}
}
\value{
The function returns one {\link[base]{data.frame}}.
}
\description{
Identifies in the observer data all the species with an inconsistent fate according to their species group.
}
\details{
The input dataframes must contain all these columns for the function to work :
\itemize{
Dataframe catch:
\item{\code{  catch_id}}
\item{\code{  speciesfate_code}}
\item{\code{  species_faocode}}
\item{\code{  speciesgroup_label1}}
}
Or
\itemize{
Dataframe sample:
\item{\code{  samplemeasure_id}}
\item{\code{  speciesfate_code}}
\item{\code{  species_faocode}}
\item{\code{  speciesgroup_label1}}
}
}
\note{
Version 1.0
}
\author{
Chloé Tellier, Philippe S. Sabarros
}
\examples{
#Catch 1 is ok: the catch is a whale and its fate is code 1,
#Catch 2 is not ok: the catch is a bycatch and its fate is code 2,
#Catch 3 is not ok: the catch is a whale and its fate is code 4,
#Catch 4 is ok: the catch is a major tuna and its fate is code 6,
#Catch 5 is not ok: the catch is a bycatch and its fate is code 6,
#Catch 6 is ok: the catch is a minor tuna and its fate is code 15,
#Catch 7 is not ok: the catch is a major tuna and is fate is code 15,
#Catch 8 is not ok: the catch is a bycatch and its fate is code 10.
catch <- data.frame(catch_id = c("1", "2", "3", "4", "5", "6", "7", "8"),
                    speciesfate_code = c("1", "2", "4", "6", "6", "15", "15", "10"),
                    species_faocode = c("MYS", "ALM", "MYS", "YFT", "CNT", "LTA", "BET", "MOP"),
                    speciesgroup_label1 = c("Cetaceans", "Other bony fishes", "Cetaceans", "Tunas nei", "Other bony fishes", "Tunas nei", "Tunas nei", "Other bony fishes"))
obs_catch_or_sample_fate_by_species_group_control(catch)
}
