% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_catch_or_sample_fate_unknown_control.R
\name{obs_catch_or_sample_fate_unknown_control}
\alias{obs_catch_or_sample_fate_unknown_control}
\title{Catch or sample fate unknown control}
\usage{
obs_catch_or_sample_fate_unknown_control(data)
}
\arguments{
\item{data}{{\link[base]{data.frame}} expected. All catches or samples during the time range selected.}
}
\value{
The function returns one {\link[base]{data.frame}}.
}
\description{
Identifies in the observer data catches or samples where fate code 9 (other) or 11 (discarded unknown) were used.
}
\details{
The input dataframes must contain all these columns for the function to work :
\itemize{
Dataframe catch:
\item{\code{  catch_id}}
\item{\code{  speciesfate_code}}
}
Or
\itemize{
Dataframe sample:
\item{\code{  samplemeasure_id}}
\item{\code{  speciesfate_code}}
}
}
\note{
Version 1.0
}
\author{
Chloé Tellier, Esther Mollier, Philippe S. Sabarros
}
\examples{
#Catch 1 is ok: the catch is discarded alive,
#Catch 2 is ok: the catch is retained on board,
#Catch 3 is not ok: the catch has an unknown fate,
#Catch 3 is not ok: the catch is discarded but with an unknown status.
catch <- data.frame(catch_id = c("1", "2", "3", "4"),
                    speciesfate_code = c("4", "15", "9", "11"))
obs_catch_or_sample_fate_unknown_control(catch)
}
