% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_catch_or_sample_outliers_control.R
\name{obs_catch_or_sample_outliers_control}
\alias{obs_catch_or_sample_outliers_control}
\title{Catch or sample outliers control}
\usage{
obs_catch_or_sample_outliers_control(data)
}
\arguments{
\item{data}{{\link[base]{data.frame}} expected. All catches or samples during the time range selected.}
}
\value{
The function returns one {\link[base]{data.frame}}.
}
\description{
Identifies for each species in the observer data the individuals with outliers lengths.
}
\details{
The input dataframes must contain all these columns for the function to work :
\itemize{
Dataframe catch:
\item{\code{  catch_id}}
\item{\code{  catch_meanlength}}
\item{\code{  catch_meanlengthcomputedsource}}
\item{\code{  species_minlength}}
\item{\code{  species_maxlength}}
}
Or
\itemize{
Dataframe sample:
\item{\code{  samplemeasure_id}}
\item{\code{  sample_length}}
\item{\code{  sample_islengthcomputed}}
\item{\code{  species_minlength}}
\item{\code{  species_maxlength}}
}
}
\note{
Version 1.0
}
\author{
Chloé Tellier, Esther Mollier, Flore Moussy, Philippe S. Sabarros
}
\examples{
#Catch 1 is ok: the catch is 30 cm, its computed source is NA, its mean length is 10 and its max length is 60,
#Catch 2 is not ok: the catch is 12 cm, its computed source is NA, its mean length is 10 and its max length is 60,
#Catch 3 is not ok: the catch is 59 cm, its computed source is NA, its mean length is 10 and its max length is 60,
#Catch 4 is not ok: the catch is 30 cm, its computed source is NA, its mean length is NA and its max length is NA.
catch <- data.frame(catch_id = c("1", "2", "3", "4"),
                    catch_meanlength = c(30, 12, 59, 30),
                    catch_meanlengthcomputedsource = c(NA_integer_, NA_integer_, NA_integer_, NA_integer_),
                    species_minlength = c(10, 10, 10, NA),
                    species_maxlength = c(60, 60, 60, NA))
obs_catch_or_sample_outliers_control(catch)
}
