% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_catch_or_sample_weight_null_control.R
\name{obs_catch_or_sample_weight_null_control}
\alias{obs_catch_or_sample_weight_null_control}
\title{Catch or sample weight unknown control}
\usage{
obs_catch_or_sample_weight_null_control(data)
}
\arguments{
\item{data}{{\link[base]{data.frame}} expected. All catches or samples during the time range selected.}
}
\value{
The function returns one {\link[base]{data.frame}}.
}
\description{
Identifies in the observer data catches or samples all individual weights = 0 or NA.
}
\details{
The input dataframes must contain all these columns for the function to work :
\itemize{
Dataframe catch:
\item{\code{  catch_id}}
\item{\code{  catch_weight}}
}
Or
\itemize{
Dataframe sample:
\item{\code{  samplemeasure_id}}
\item{\code{  sample_weight}}
}
}
\note{
Version 1.0
}
\author{
Chloé Tellier, Philippe S. Sabarros
}
\examples{
#Catch 1 is ok: the catch's weight is > 0,
#Catch 2 is not ok: the catch's weight is = 0,
#Catch 3 is not ok: the catch's weight is NA.
catch <- data.frame(catch_id = c("1", "2", "3"),
                    catch_weight = c("5", "0", NA))
obs_catch_or_sample_weight_null_control(catch)
}
