% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_sample_measure_type_control.R
\name{obs_sample_measure_type_control}
\alias{obs_sample_measure_type_control}
\title{Sample measure type control}
\usage{
obs_sample_measure_type_control(sample)
}
\arguments{
\item{sample}{{\link[base]{data.frame}} expected. All samples during the time range selected.}
}
\value{
The function returns one {\link[base]{data.frame}}.
}
\description{
This function allows to check the coherence of a sampled species measure type according to its default measure type. Major tunas (YFT, BET, SKJ and ALB) could also be measured in PD1.
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe sample:
\item{\code{  samplemeasure_id}}
\item{\code{  species_faocode}}
\item{\code{  sizetype_code}}
\item{\code{  defaultsizetype_code}}
}
}
\note{
Version 1.0
}
\author{
Chloé Tellier, Esther Mollier, Philippe S. Sabarros
}
\examples{
#Sample 1 is ok: the sample is a CNT measured in TL
#Sample 2 is ok: the sample is a FAL measured in TL
#Sample 3 is ok: the sample is a LKV measured in SCL
#Sample 4 is ok: the sample is a YFT measured in PD1
#Sample 5 is not ok: the sample is a FAL measured in FL
sample <- data.frame(samplemeasure_id = c("1", "2", "3", "4", "5"),
                     species_faocode = c("CNT", "FAL", "LKV", "YFT", "FAL"),
                     sizetype_code = c("TL", "TL", "SCL", "PD1", "FL"),
                     defaultsizetype_code = c("TL", "TL", "SCL", "FL", "TL"))
obs_sample_measure_type_control(sample)
}
