% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_sample_more_than_catch_control.R
\name{obs_sample_more_than_catch_control}
\alias{obs_sample_more_than_catch_control}
\title{Sample more than catch control}
\usage{
obs_sample_more_than_catch_control(catch, sample)
}
\arguments{
\item{catch}{{\link[base]{data.frame}} expected. All catches during the time range selected.}

\item{sample}{{\link[base]{data.frame}} expected. All samples during the time range selected.}
}
\value{
The function returns one {\link[base]{data.frame}}.
}
\description{
Identifies in the observer data all the observations with a number of samples superior to the number of catches.
}
\details{
The input dataframes must contain all these columns for the function to work :
\itemize{
Dataframe catch:
\item{\code{  catch_id}}
\item{\code{  activity_id}}
\item{\code{  species_faocode}}
\item{\code{  speciesfate_code}}
\item{\code{  catch_count}}
}
\itemize{
Dataframe sample:
\item{\code{  samplemeasure_id}}
\item{\code{  activity_id}}
\item{\code{  species_faocode}}
\item{\code{  speciesfate_code}}
\item{\code{  sample_count}}
}
}
\note{
Version 1.0
}
\author{
Chloé Tellier, Esther Mollier, Philippe S. Sabarros
}
\examples{
#Activity 1 is ok: there is 2 catches and 1 samples,
#Activity 2 is ok: there is 2 catches and 2 samples,
#Activity 3 is not ok: there is 1 catch but 2 samples.
catch <- data.frame(catch_id = c("1", "2", "3", "4"),
                    activity_id = c("1", "2", "2", "3"),
                    species_faocode = c("BLT", "FAL", "FAL", "CNT"),
                    speciesfate_code = c("15", "4", "4", "5"),
                    catch_count = c("2", "1", "1", "1"))
sample <- data.frame(samplemeasure_id = c("1", "2", "3", "4"),
                     activity_id = c("1", "2", "2", "3"),
                     species_faocode = c("BLT", "FAL", "FAL", "CNT"),
                     speciesfate_code = c("15", "4", "4", "5"),
                     sample_count = c("1", "1", "1", "2"))
obs_sample_more_than_catch_control(catch, sample)
}
