% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_sample_no_fate_control.R
\name{obs_sample_no_fate_control}
\alias{obs_sample_no_fate_control}
\title{Sample with no fate control}
\usage{
obs_sample_no_fate_control(sample)
}
\arguments{
\item{sample}{{\link[base]{data.frame}} expected. All samples during the time range selected.}
}
\value{
The function returns one {\link[base]{data.frame}}.
}
\description{
This function aims to find all the samples in the observer data that don't have a fate.
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe sample:
\item{\code{  samplemeasure_id}}
\item{\code{  speciesfate_code}}
}
}
\note{
Version 1.0
}
\author{
Chloé Tellier, Esther Mollier, Philippe S. Sabarros
}
\examples{
#Sample 1 is ok: the fate code is 6 (retained on board)
#Sample 2 is ok: the fate code is 4 (discarded alive)
#Sample 3 is not ok: the fate code is NA
sample <- data.frame(samplemeasure_id = c("1", "2", "3"),
                     speciesfate_code = c("6", "4", NA))
obs_sample_no_fate_control(sample)
}
