% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observer_data_quality_control.R
\name{observer_data_quality_control}
\alias{observer_data_quality_control}
\title{observer data quality control}
\usage{
observer_data_quality_control(
  data_connection,
  start_year,
  end_year,
  program,
  ocean,
  country_code,
  observer_first_name,
  path_file = NULL
)
}
\arguments{
\item{data_connection}{{\link[base]{list}} expected. Either output of the function {\link[furdeb]{postgresql_dbconnection}}, which must be done before using the size_distribution_and_outliers_control}

\item{start_year}{{\link[base]{integer}} expected. Starting year for the samples to control.}

\item{end_year}{{\link[base]{integer}} expected. Ending year for the samples to control.}

\item{program}{{\link[base]{character}} expected. Programs to be controlled. Example of the format for a program topiaid: "fr.ird.referential.ps.common.Program#1239832686262#0.31033946454061234"}

\item{ocean}{{\link[base]{character}} expected. Ocean to be controlled. Examples: 'Indian', 'Atlantic'...etc.}

\item{country_code}{{\link[base]{character}} expected. Countries on wich control will be made. Examples: 'FRA', 'MUS'...etc.}

\item{observer_first_name}{{\link[base]{character}} expected. Observer first name we want to control in priority. Examples: 'Martin', 'Durand'...etc.}

\item{path_file}{{\link[base]{character}} expected. Path to save the final xlsx.}
}
\value{
The function return histograms and xlsx tables.
}
\description{
'This script aims to generate indicators for evaluating the quality of data collected by observers'
}
