% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_table_checking.R
\name{r_table_checking}
\alias{r_table_checking}
\title{Checking for table object}
\usage{
r_table_checking(
  r_table,
  type,
  column_name = NULL,
  column_type = NULL,
  output = "error"
)
}
\arguments{
\item{r_table}{Table object expected.}

\item{type}{Object of class \code{\link[base]{character}} expected. Type or class expected of the r table You can choose between arguments "data.frame", or "tbl".}

\item{column_name}{\code{\link[base]{vector}} of {\link[base]{character}} expected. Be default NULL. Mandatory column names in r_table.}

\item{column_type}{\code{\link[base]{vector}} of {\link[base]{character}} expected. Be default NULL. Mandatory column type in r_table.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. By default "error". You can choose between "error", "message" or "logical".}
}
\value{
The function returns error with output is "error" and if the parameters are not respected, a {\link[base]{character}} with output is "message", a {\link[base]{logical}} with output is "logical"
}
\description{
Type, name and type of mandatory columns are controlled
}
