% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_type_checking.R
\name{r_type_checking}
\alias{r_type_checking}
\title{Checking for r object type}
\usage{
r_type_checking(r_object, type, length, allowed_values = NULL, silent = TRUE)
}
\arguments{
\item{r_object}{R object expected.}

\item{type}{Object of class \code{\link[base]{character}} expected. Type or class expected of the r object. you can choose between arguments "character", "list", "numeric", "integer" or "logical".}

\item{length}{Object of class \code{\link[base]{integer}} expected. Specify the vector length expected.}

\item{allowed_values}{Object of class \code{\link[base]{vector}} (excepted a list) expected. Be default NULL. Allowed value(s) in the r object.}

\item{silent}{Object of class \code{\link[base]{logical}} expected. By default TRUE. Display output message when process ran  Display output message when process ran  Display output message when process ran sucessfully.}
}
\description{
Typing and structure checking for r object.
}
