% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_control.R
\name{sex_control}
\alias{sex_control}
\title{Sex control}
\usage{
sex_control(
  data_connection,
  start_year,
  end_year,
  program,
  ocean,
  country_code,
  path_file = NULL,
  export_table
)
}
\arguments{
\item{data_connection}{{\link[base]{list}} expected. Either output of the function {\link[furdeb]{postgresql_dbconnection}}, which must be done before using the sex_control.}

\item{start_year}{{\link[base]{integer}} expected. Starting year for the control.}

\item{end_year}{{\link[base]{integer}} expected. Ending year for the control.}

\item{program}{{\link[base]{character}} expected. Programs to be controlled. Example of the format for a program topiaid: "fr.ird.referential.ps.common.Program#1239832686262#0.31033946454061234".}

\item{ocean}{{\link[base]{character}} expected. Ocean to be controlled. Examples: 'Indian', 'Atlantic'...etc.}

\item{country_code}{{\link[base]{character}} expected. Countries on which control will be made. Examples: 'FRA', 'MUS'...etc.}

\item{path_file}{{\link[base]{character}} expected. By default NULL. Path to save the final xlsx.}

\item{export_table}{{\link[base]{logical}} expected. Save or not the xlsx table.}
}
\value{
The function returns one xlsx table.
}
\description{
This function aims to check the coherence of the sex of a sample according to the possibility of sexing this species. For each inconsistency, it returns a table containing the details of the observation to correct.
}
\note{
Version 1.0
}
\author{
Chloé Tellier, Philippe S. Sabarros
}
