% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_comparison.R
\name{vector_comparison}
\alias{vector_comparison}
\title{Vector comparison}
\usage{
vector_comparison(
  first_vector,
  second_vector,
  comparison_type,
  output,
  epsilon = 1.5e-08
)
}
\arguments{
\item{first_vector}{{\link[base]{vector}} expected.}

\item{second_vector}{{\link[base]{vector}} expected.}

\item{comparison_type}{{\link[base]{character}} expected. Type of comparison expected, you can choose between "difference", "equal", "less", "greater", "less_equal" or "greater_equal". "Difference" highlight element(s) of the first vector not present in the second, "equal" check if two vectors are exactly the same, "less" check if the elements of the first vector are strictly less than their pair in the second vector , "greater" check if the elements of the first vector are strictly greater than their pair in the second vector, "less_equal" check if the elements of the first vector are less than or equal to their pair in the second vector or "greater_equal" check if the elements of the first vector are greater than or equal to their pair in the second vector.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{epsilon}{{\link[base]{numeric}} expected. Default values: 1.5e-8. Tolerance of comparison for numeric value, following the problem caused by IEEE Standard 754}
}
\description{
Identify similarities and differences between two R vectors (for the numbers the comparison is not exact, an epsilon is used).
}
\examples{
vector_comparison(
  first_vector = c(1, 2, 3, 5),
  second_vector = c(1, 2, 4, 3),
  comparison_type = "difference",
  output = "report"
)
}
