% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors_comparisons.R
\name{vectors_comparisons}
\alias{vectors_comparisons}
\title{Vectors comparisons}
\usage{
vectors_comparisons(first_vector, second_vector, comparison_type, output)
}
\arguments{
\item{first_vector}{{\link[base]{vector}} expected.}

\item{second_vector}{{\link[base]{vector}} expected.}

\item{comparison_type}{{\link[base]{character}} expected. Type of comparison expected, you can choose between "difference" or "equality". "Difference" highlight element(s) of the first vector not present in the second, and "equality" check if two vectors are exactly the same.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}
}
\description{
Identify similarities and differences between two R vectors
}
