% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codama-package.R
\docType{package}
\name{codama-package}
\alias{codama}
\alias{codama-package}
\title{Common prOcesses for Data quAlity Monitoring}
\description{
This package aims is to provide a series of standardised and generic controls processes for data quality assessment and monitoring.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OB7-IRD/codama}
  \item Report bugs at \url{https://github.com/OB7-IRD/codama/issues}
}

}
\author{
\strong{Maintainer}: Ob7 - Observatory of Exploited Tropical Pelagic Ecosystems \email{ob7@ird.fr} [copyright holder]

Authors:
\itemize{
  \item Mathieu Depetris \email{mathieu.depetris@ird.fr}
}

}
\keyword{internal}
