% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codama-package.R
\docType{package}
\name{codama-package}
\alias{codama}
\alias{codama-package}
\title{Common prOcesses for Data quAlity Monitoring}
\description{
This package aims is to provide a series of standardised and generic controls processes for data quality assessment and monitoring.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OB7-IRD/codama}
  \item Report bugs at \url{https://github.com/OB7-IRD/codama/issues}
}

}
\author{
\strong{Maintainer}: Oceane Bouhineau \email{oceane.bouhineau@ird.fr} (\href{https://orcid.org/0000-0001-7418-2925}{ORCID})

Authors:
\itemize{
  \item Mathieu Depetris \email{mathieu.depetris@ird.fr} (\href{https://orcid.org/0000-0001-8080-0531}{ORCID})
  \item Theotime Fily \email{theotime.fily@ird.fr} (\href{https://orcid.org/0000-0001-7277-1769}{ORCID})
  \item Julien Lebranchu \email{julien.lebranchu@ird.fr} (\href{https://orcid.org/0000-0002-7759-0534}{ORCID})
  \item Clara Lerebourg \email{clara.lerebourg@ird.fr} (\href{https://orcid.org/0000-0003-0589-8932}{ORCID})
  \item Esther Mollier \email{esther.mollier@ird.fr} (\href{https://orcid.org/0000-0001-8921-079X}{ORCID})
  \item Chloé Tellier \email{chloe.tellier@ird.fr} (\href{https://orcid.org/0009-0007-5988-3310}{ORCID})
  \item Philippe Sabarros \email{philippe.sabarros@ird.fr} (\href{https://orcid.org/0000-0001-6135-9015}{ORCID})
}

Other contributors:
\itemize{
  \item IRD - Ob7 Observatory of Exploited Tropical Pelagic Ecosystems \email{ob7@ird.fr} (\href{https://orcid.org/0000-0002-4642-3418}{ORCID}) [copyright holder, funder]
}

}
\keyword{internal}
