% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_path_checking.R
\name{file_path_checking}
\alias{file_path_checking}
\title{Checking for file path}
\usage{
file_path_checking(file_path, extension, silent = TRUE)
}
\arguments{
\item{file_path}{Object of class \code{\link[base]{character}} expected. File location on your system.}

\item{extension}{Object of class \code{\link[base]{character}} expected. Type of extension expected (avoid the . before the extension).}

\item{silent}{Object of class \code{\link[base]{logical}} expected. By default TRUE. Display output message when process ran  Display output message when process ran  Display output message when process ran sucessfully.}
}
\description{
Checking for file path before import in R.
}
