% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/logbook_vessel_activity_floating_object_missing_buoy_control.R
\name{logbook_vessel_activity_floating_object_missing_buoy_control}
\alias{logbook_vessel_activity_floating_object_missing_buoy_control}
\title{Identification of incoherent vessel activity and missing buoy}
\usage{
logbook_vessel_activity_floating_object_missing_buoy_control(
  dataframe1,
  dataframe2,
  dataframe3,
  output,
  vessel_activity = c("13")
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_vessel_activity_floating_object_missing_buoy_control.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_vessel_activity_floating_object_missing_buoy_control.}

\item{dataframe3}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the logbook_vessel_activity_floating_object_missing_buoy_control.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{vessel_activity}{{\link[base]{character}} expected. Default values: c("13"). Vector containing the vessel activity code that must have a buoy.}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical".
}
\description{
The purpose of the logbook_vessel_activity_floating_object_missing_buoy_control function is to provide a table of data that contains a incoherent vessel activity and missing buoy.
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
\item{\code{  floatingobject_id}}
\item{\code{  activity_id}}
}
\itemize{
Dataframe 2:
\item{\code{  activity_id}}
\item{\code{  vesselactivity_code}}
}
\itemize{
Dataframe 3:
\item{\code{  transmittingbuoy_id}}
\item{\code{  floatingobject_id}}
}
}
\examples{
#Floating object 1, 2 and 3 are ok,
#Floating object 4 has buoy missing
#Floating object 5 has vessel activity missing
dataframe1 <- data.frame(floatingobject_id = c("1", "2", "3", "4", "5"),
                         activity_id = c("1", "2", "3", "3", "4"))
dataframe2 <- data.frame(activity_id = c("1", "2", "3", "4"),
                         vesselactivity_code = c("13", "1", "13", NA))
dataframe3 <- data.frame(transmittingbuoy_id = c("1", "2"),
                         floatingobject_id = c("1", "3"))
logbook_vessel_activity_floating_object_missing_buoy_control(dataframe1,
                                                             dataframe2,
                                                             dataframe3,
                                                             output = "report")
}
