% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_activity_observed_systems_control.R
\name{obs_activity_observed_systems_control}
\alias{obs_activity_observed_systems_control}
\title{Activity observed systems control}
\usage{
obs_activity_observed_systems_control(catch, observedsystem)
}
\arguments{
\item{catch}{{\link[base]{data.frame}} expected. All catches during the time range selected.}

\item{observedsystem}{{\link[base]{data.frame}} expected. All observed systems for activities during the time range selected.}
}
\value{
The function returns one {\link[base]{data.frame}}.
}
\description{
Identifies in the observer data any inconsistencies between observed systems and species caught.
}
\details{
The input dataframes must contain all these columns for the function to work :
\itemize{
Dataframe catch:
\item{\code{  activity_id}}
\item{\code{  catch_id}}
\item{\code{  species_faocode}}
}
\itemize{
Dataframe observedsystem:
\item{\code{  activity_id}}
\item{\code{  observedsystem_id}}
\item{\code{  observedsystem_code}}
}
}
\note{
Version 1.0
}
\author{
Chloé Tellier, Philippe S. Sabarros
}
\examples{
#Catch 1 is ok: there is a whale shark in the catches, and a whale shark in the observed systems.
#Catch 2 is ok: there is a whale in the catches, and a whale in the observed systems.
#Catch 3 is ok: there is no whale or whale shark in the catches, and no whale or whale shark in the observed systems.
#Catch 4 is not ok: there is a whale in the catches, but not in the observed systems.
#Catch 5 is not ok: there is a whale shark in the observed systems, but not in the catches.
catch <- data.frame(activity_id = c("1", "2", "3", "4", "5"),
                    catch_id = c("1", "2", "3", "4", "5"),
                    species_faocode = c("RHN, SKJ, BLT", "MYS, YFT, BET", "SKJ, YFT", "SKJ, MYS, BUM", "YFT"))
observedsystem <- data.frame(activity_id = c("1", "2", "3", "4", "5"),
                             observedsystem_id = c("1", "2", "3", "4", "5"),
                             observedsystem_code = c("22", "112", "0", "0", "21"))
obs_activity_observed_systems_control(catch, observedsystem)
}
