% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_sample_round_size_control.R
\name{obs_sample_round_size_control}
\alias{obs_sample_round_size_control}
\title{Sample round size control}
\usage{
obs_sample_round_size_control(sample)
}
\arguments{
\item{sample}{{\link[base]{data.frame}} expected. All samples during the time range selected.}
}
\value{
The function returns one {\link[base]{data.frame}}.
}
\description{
This function controls the accuracy (decimals) of size samples. Returns samples with decimal values that are not rounded to the inferior centimeter (or to the nearest half centimeter for PD1 measurements).
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe sample:
\item{\code{  samplemeasure_id}}
\item{\code{  sample_length}}
\item{\code{  sample_islengthcomputed}}
\item{\code{  sizetype_code}}
}
}
\note{
Version 1.0
}
\author{
Chloé Tellier, Philippe S. Sabarros
}
\examples{
#Sample 1 is ok: the size type is TL and the length is rounded to .0
#Sample 2 is ok: the size type is PD1 and the length is rounded to.5
#Sample 3 is not ok: the size type is TL and the length is not rounded to .0
#Sample 4 is not ok: the size type is PD1 and the length is not rounded to .5
sample <- data.frame(samplemeasure_id = c("1", "2", "3", "4"),
                     sample_length = c("112", "35.5", "119.3", "43.7"),
                     sample_islengthcomputed = c("FALSE", "FALSE", "FALSE", "FALSE"),
                     sizetype_code = c("TL", "PD1", "TL", "PD1"))
obs_sample_round_size_control(sample)
}
