% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_sample_sex_control.R
\name{obs_sample_sex_control}
\alias{obs_sample_sex_control}
\title{Sex control}
\usage{
obs_sample_sex_control(sample)
}
\arguments{
\item{sample}{{\link[base]{data.frame}} expected. All samples during the time range selected.}
}
\value{
The function returns one {\link[base]{data.frame}}.
}
\description{
This function aims to check the coherence of the sex of a sample according to the possibility of sexing this species. For each inconsistency, it returns a table containing the details of the observation to correct.
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe sample:
\item{\code{  samplemeasure_id}}
\item{\code{  sex_label}}
\item{\code{  species_faocode}}
\item{\code{  speciesgroup_label}}
}
}
\note{
Version 1.0
}
\author{
Chloé Tellier, Philippe S. Sabarros
}
\examples{
#Sample 1 is ok: the sample is not sexed
#Sample 2 is ok: the sample is sexed and is a DOL
#Sample 3 is ok: the sample is sexed and is a shark
#Sample 4 is not ok: the sample is sexed and is a CNT
sample <- data.frame(samplemeasure_id = c("1", "2", "3", "4"),
                     sex_label = c(NA, "Female", "Female", "Male"),
                     species_faocode = c("CNT", "DOL", "FAL", "CNT"),
                     speciesgroup_label = c("Other bony fishes",
                                            "Other bony fishes",
                                            "Sharks",
                                            "Other bony fishes"))
obs_sample_sex_control(sample)
}
