% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_type_checking.R
\name{r_type_checking}
\alias{r_type_checking}
\title{Checking for r object type}
\usage{
r_type_checking(
  r_object,
  type,
  length = NULL,
  allowed_value = NULL,
  output = "error"
)
}
\arguments{
\item{r_object}{R object expected.}

\item{type}{Object of class \code{\link[base]{character}} expected. Type or class expected of the r object. you can choose between arguments "character", "list", "PostgreSQLConnection", "PqConnection", "numeric", "integer", "logical", "data.frame", "tbl" or "NULL".}

\item{length}{Object of class \code{\link[base]{integer}} expected. By default NULL. Specify the vector length expected. Fill with NULL with you don't know the length expected.}

\item{allowed_value}{Object of class \code{\link[base]{vector}} (excepted a list) expected. Be default NULL. Allowed value(s) in the r object.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. By default "error". You can choose between "error", "message" or "logical".}
}
\value{
The function returns error with output is "error" and if the parameters are not respected, a {\link[base]{character}} with output is "message", a {\link[base]{logical}} with output is "logical"
}
\description{
Typing and structure checking for r object.
}
